DiggleGratton {spatstat}	R Documentation

Diggle-Gratton model

Description

Creates an instance of the Diggle-Gratton pairwise interaction point process model, which can then be fitted to point pattern data.

Usage


  DiggleGratton(delta, rho)

Arguments

delta 	 lower threshold delta

rho 	 upper threshold rho

Details

Diggle and Gratton (1984, pages 208-210) introduced the pairwise interaction point process with pair potential h(t) of the form

h(t) = ((t - delta)/(rho - delta))^kappa, { } delta <= t <= rho

with h(t) = 0 for t < delta and h(t) = 1 for t > rho. Here delta, rho and kappa are parameters.

Note that we use the symbol kappa where Diggle and Gratton (1984) and Diggle, Gates and Stibbard (1987) use beta, since in spatstat we reserve the symbol beta for an intensity parameter.

The parameters must all be nonnegative, and must satisfy delta <= rho.

The potential is inhibitory, i.e. this model is only appropriate for regular point patterns. The strength of inhibition increases with kappa. For kappa=0 the model is a hard core process with hard core radius delta. For kappa=infinity the model is a hard core process with hard core radius rho.

The irregular parameters delta, rho must be given in the call to DiggleGratton, while the regular parameter kappa will be estimated.

Value

An object of class "interact" describing the interpoint interaction structure of a point process.

Author(s)

Adrian Baddeley adrian@maths.uwa.edu.au http://www.maths.uwa.edu.au/~adrian/ and Rolf Turner rolf@math.unb.ca http://www.math.unb.ca/~rolf

References

Diggle, P.J., Gates, D.J. and Stibbard, A. (1987) A nonparametric estimator for pairwise-interaction point processes. Biometrika 74, 763 – 770.

Diggle, P.J. and Gratton, R.J. (1984) Monte Carlo methods of inference for implicit statistical models. Journal of the Royal Statistical Society, series B 46, 193 – 212.

See Also

ppm, ppm.object, Pairwise

Examples


   data(cells)
   ppm(cells, ~1, DiggleGratton(0.05, 0.1))

[Package spatstat version 1.10-3 Index]