Gdot {spatstat}	R Documentation

Multitype Nearest Neighbour Distance Function (i-to-any)

Description

For a multitype point pattern, estimate the distribution of the distance from a point of type i to the nearest other point of any type.

Usage


Gdot(X, i=1, r=NULL, breaks=NULL, ...)

Arguments

X 	 The observed point pattern, from which an estimate of the distance distribution function Gi.(r) will be computed. It must be a multitype point pattern (a marked point pattern whose marks are a factor). See under Details.

i 	 Number or character string identifying the type (mark value) of the points in X from which distances are measured.

r 	 Optional. Numeric vector. The values of the argument r at which the distribution function Gi.(r) should be evaluated. There is a sensible default. First-time users are strongly advised not to specify this argument. See below for important conditions on r.

breaks 	 An alternative to the argument r. Not normally invoked by the user. See the Details section.

... 	 Ignored.

Details

This function Gdot and its companions Gcross and Gmulti are generalisations of the function Gest to multitype point patterns.

A multitype point pattern is a spatial pattern of points classified into a finite number of possible ``colours'' or ``types''. In the spatstat package, a multitype pattern is represented as a single point pattern object in which the points carry marks, and the mark value attached to each point determines the type of that point.

The argument X must be a point pattern (object of class "ppp") or any data that are acceptable to as.ppp. It must be a marked point pattern, and the mark vector X$marks must be a factor. The argument will be interpreted as a level of the factor X$marks. (Warning: this means that an integer value i=3 will be interpreted as the 3rd smallest level, not the number 3).

The ``dot-type'' (type i to any type) nearest neighbour distance distribution function of a multitype point process is the cumulative distribution function Gi.(r) of the distance from a typical random point of the process with type i the nearest other point of the process, regardless of type.

An estimate of Gi.(r) is a useful summary statistic in exploratory data analysis of a multitype point pattern. If the type i points were independent of all other points, then Gi.(r) would equal Gii(r), the nearest neighbour distance distribution function of the type i points alone. For a multitype Poisson point process with total intensity lambda, we have

Gi.(r) = 1 - exp( - lambda * pi * r^2)

Deviations between the empirical and theoretical Gi. curves may suggest dependence of the type i points on the other points.

This algorithm estimates the distribution function Gi.(r) from the point pattern X. It assumes that X can be treated as a realisation of a stationary (spatially homogeneous) random spatial point process in the plane, observed through a bounded window. The window (which is specified in X as X$window) may have arbitrary shape. Biases due to edge effects are treated in the same manner as in Gest.

The argument r is the vector of values for the distance r at which Gi.(r) should be evaluated. It is also used to determine the breakpoints (in the sense of hist) for the computation of histograms of distances. The reduced-sample and Kaplan-Meier estimators are computed from histogram counts. In the case of the Kaplan-Meier estimator this introduces a discretisation error which is controlled by the fineness of the breakpoints.

First-time users would be strongly advised not to specify r. However, if it is specified, r must satisfy r[1] = 0, and max(r) must be larger than the radius of the largest disc contained in the window. Furthermore, the successive entries of r must be finely spaced.

The algorithm also returns an estimate of the hazard rate function, lambda(r), of Gi.(r). This estimate should be used with caution as Gi.(r) is not necessarily differentiable.

The naive empirical distribution of distances from each point of the pattern X to the nearest other point of the pattern, is a biased estimate of Gi.. However this is also returned by the algorithm, as it is sometimes useful in other contexts. Care should be taken not to use the uncorrected empirical Gi. as if it were an unbiased estimator of Gi..

Value

An object of class "fv" (see fv.object).
Essentially a data frame containing six numeric columns

r 	 the values of the argument r at which the function Gi.(r) has been estimated

rs 	 the ``reduced sample'' or ``border correction'' estimator of Gi.(r)

km 	 the spatial Kaplan-Meier estimator of Gi.(r)

hazard 	 the hazard rate lambda(r) of Gi.(r) by the spatial Kaplan-Meier method

raw 	 the uncorrected estimate of Gi.(r), i.e. the empirical distribution of the distances from each point of type i to the nearest other point of any type.

theo 	 the theoretical value of Gi.(r) for a marked Poisson process with the same estimated intensity (see below).

Warnings

The argument i is interpreted as a level of the factor X$marks. Beware of the usual trap with factors: numerical values are not interpreted in the same way as character values. See the first example.

The function Gi. does not necessarily have a density.

The reduced sample estimator of Gi. is pointwise approximately unbiased, but need not be a valid distribution function; it may not be a nondecreasing function of r. Its range is always within [0,1].

The spatial Kaplan-Meier estimator of Gi. is always nondecreasing but its maximum value may be less than 1.

Author(s)

Adrian Baddeley adrian@maths.uwa.edu.au http://www.maths.uwa.edu.au/~adrian/ and Rolf Turner rolf@math.unb.ca http://www.math.unb.ca/~rolf

References

Cressie, N.A.C. Statistics for spatial data. John Wiley and Sons, 1991.

Diggle, P.J. Statistical analysis of spatial point patterns. Academic Press, 1983.

Diggle, P. J. (1986). Displaced amacrine cells in the retina of a rabbit : analysis of a bivariate spatial point pattern. J. Neurosci. Meth. 18, 115–125.

Harkness, R.D and Isham, V. (1983) A bivariate spatial point pattern of ants' nests. Applied Statistics 32, 293–303

Lotwick, H. W. and Silverman, B. W. (1982). Methods for analysing spatial processes of several types of points. J. Royal Statist. Soc. Ser. B 44, 406–413.

Ripley, B.D. Statistical inference for spatial processes. Cambridge University Press, 1988.

Stoyan, D, Kendall, W.S. and Mecke, J. Stochastic geometry and its applications. 2nd edition. Springer Verlag, 1995.

Van Lieshout, M.N.M. and Baddeley, A.J. (1999) Indices of dependence between types in multivariate point patterns. Scandinavian Journal of Statistics 26, 511–532.

See Also

Gcross, Gest, Gmulti

Examples


    data(betacells)
     # cat retina data
    G0. <- Gdot(betacells, "off") 
    plot(G0.)

    # synthetic example    
    pp <- runifpoispp(50)
    pp <- pp %mark% factor(sample(0:1, pp$n, replace=TRUE))
    G <- Gdot(pp, "0") # note: "0" not 0

[Package spatstat version 1.10-3 Index]