Geyer {spatstat}	R Documentation

Geyer's Saturation Point Process Model

Description

Creates an instance of Geyer's ``saturation point process'' model which can then be fitted to point pattern data.

Usage


  Geyer(r,sat)

Arguments

r 	 Interaction radius. A positive real number.

sat 	 Saturation threshold. A positive real number.

Details

Geyer (1999) introduced the ``saturation process'', a modification of the Strauss process (see Strauss) in which the total contribution to the potential from each point (from its pairwise interaction with all other points) is trimmed to a maximum value s. This model is implemented in the function Geyer().

The saturation point process with interaction radius r, saturation threshold s, and parameters beta and gamma, is the point process in which each point x[i] in the pattern X contributes a factor

beta gamma^min(s, t(x[i],X))

to the probability density of the point pattern, where t(x[i],X) denotes the number of ``close neighbours'' of x[i] in the pattern X. A close neighbour of x[i] is a point x[j] with j != i such that the distance between x[i] and x[j] is less than or equal to r.

If the saturation threshold s is set to infinity, this model reduces to the Strauss process (see Strauss) with interaction parameter gamma^2. If s = 0, the model reduces to the Poisson point process. If s is a finite positive number, then the interaction parameter gamma may take any positive value (unlike the case of the Strauss process), with values gamma < 1 describing an ``ordered'' or ``inhibitive'' pattern, and values gamma > 1 describing a ``clustered or ``attractive'' pattern.

The nonstationary saturation process is similar except that the value beta is replaced by a function beta(x[i]) of location.

The function ppm(), which fits point process models to point pattern data, requires an argument of class "interact" describing the interpoint interaction structure of the model to be fitted. The appropriate description of the saturation process interaction is yielded by Geyer(r, sat) where the arguments r and sat specify the Strauss interaction radius r and the saturation threshold s, respectively. See the examples below.

Note the only arguments are the interaction radius r and the saturation threshold sat. When r and sat are fixed, the model becomes an exponential family. The canonical parameters log(beta) and log(gamma) are estimated by ppm(), not fixed in Geyer().

Value

An object of class "interact" describing the interpoint interaction structure of Geyer's ``saturation point process'' with interaction radius r and saturation threshold sat.

Author(s)

Adrian Baddeley adrian@maths.uwa.edu.au http://www.maths.uwa.edu.au/~adrian/ and Rolf Turner rolf@math.unb.ca http://www.math.unb.ca/~rolf

References

Geyer, C.J. (1999) Likelihood Inference for Spatial Point Processes. Chapter 3 in O.E. Barndorff-Nielsen, W.S. Kendall and M.N.M. Van Lieshout (eds) Stochastic Geometry: Likelihood and Computation, Chapman and Hall / CRC, Monographs on Statistics and Applied Probability, number 80. Pages 79–140.

See Also

ppm, pairwise.family, ppm.object, Strauss

Examples


   data(cells) 
   ppm(cells, ~1, Geyer(r=0.07, sat=2), rbord=0.07)
   # fit the stationary saturation process to `cells'

[Package spatstat version 1.10-3 Index]