Gmulti {spatstat}	R Documentation

Marked Nearest Neighbour Distance Function

Description

For a marked point pattern, estimate the distribution of the distance from a typical point in subset I to the nearest point of subset J.

Usage


Gmulti(X, I, J, r=NULL, breaks=NULL, ..., disjoint=NULL)

Arguments

X 	 The observed point pattern, from which an estimate of the multitype distance distribution function GIJ(r) will be computed. It must be a marked point pattern. See under Details.

I 	 Subset of points of X from which distances are measured.

J 	 Subset of points in X to which distances are measured.

r 	 Optional. Numeric vector. The values of the argument r at which the distribution function GIJ(r) should be evaluated. There is a sensible default. First-time users are strongly advised not to specify this argument. See below for important conditions on r.

breaks 	 An alternative to the argument r. Not normally invoked by the user. See the Details section.

... 	 Ignored.

disjoint 	 Optional flag indicating whether the subsets I and J are disjoint. If missing, this value will be computed by inspecting the vectors I and J

.

Details

The function Gmulti generalises Gest (for unmarked point patterns) and Gdot and Gcross (for multitype point patterns) to arbitrary marked point patterns.

Suppose X[I], X[J] are subsets, possibly overlapping, of a marked point process. This function computes an estimate of the cumulative distribution function GIJ(r) of the distance from a typical point of X[I] to the nearest distinct point of X[J].

The argument X must be a point pattern (object of class "ppp") or any data that are acceptable to as.ppp.

The arguments I and J specify two subsets of the point pattern. They may be logical vectors of length equal to X$n, or integer vectors with entries in the range 1 to X$n, etc.

This algorithm estimates the distribution function GIJ(r) from the point pattern X. It assumes that X can be treated as a realisation of a stationary (spatially homogeneous) random spatial point process in the plane, observed through a bounded window. The window (which is specified in X as X$window) may have arbitrary shape. Biases due to edge effects are treated in the same manner as in Gest.

The argument r is the vector of values for the distance r at which GIJ(r) should be evaluated. It is also used to determine the breakpoints (in the sense of hist) for the computation of histograms of distances. The reduced-sample and Kaplan-Meier estimators are computed from histogram counts. In the case of the Kaplan-Meier estimator this introduces a discretisation error which is controlled by the fineness of the breakpoints.

First-time users would be strongly advised not to specify r. However, if it is specified, r must satisfy r[1] = 0, and max(r) must be larger than the radius of the largest disc contained in the window. Furthermore, the successive entries of r must be finely spaced.

The algorithm also returns an estimate of the hazard rate function, lambda(r), of GIJ(r). This estimate should be used with caution as GIJ(r) is not necessarily differentiable.

The naive empirical distribution of distances from each point of the pattern X to the nearest other point of the pattern, is a biased estimate of GIJ. However this is also returned by the algorithm, as it is sometimes useful in other contexts. Care should be taken not to use the uncorrected empirical GIJ as if it were an unbiased estimator of GIJ.

Value

An object of class "fv" (see fv.object).
Essentially a data frame containing six numeric columns

r 	 the values of the argument r at which the function GIJ(r) has been estimated

rs 	 the ``reduced sample'' or ``border correction'' estimator of GIJ(r)

km 	 the spatial Kaplan-Meier estimator of GIJ(r)

hazard 	 the hazard rate lambda(r) of GIJ(r) by the spatial Kaplan-Meier method

raw 	 the uncorrected estimate of GIJ(r), i.e. the empirical distribution of the distances from each point of type i to the nearest point of type j

theo 	 the theoretical value of GIJ(r) for a marked Poisson process with the same estimated intensity

Warnings

The function GIJ does not necessarily have a density.

The reduced sample estimator of GIJ is pointwise approximately unbiased, but need not be a valid distribution function; it may not be a nondecreasing function of r. Its range is always within [0,1].

The spatial Kaplan-Meier estimator of GIJ is always nondecreasing but its maximum value may be less than 1.

Author(s)

Adrian Baddeley adrian@maths.uwa.edu.au http://www.maths.uwa.edu.au/~adrian/ and Rolf Turner rolf@math.unb.ca http://www.math.unb.ca/~rolf

References

Cressie, N.A.C. Statistics for spatial data. John Wiley and Sons, 1991.

Diggle, P.J. Statistical analysis of spatial point patterns. Academic Press, 1983.

Diggle, P. J. (1986). Displaced amacrine cells in the retina of a rabbit : analysis of a bivariate spatial point pattern. J. Neurosci. Meth. 18, 115–125.

Harkness, R.D and Isham, V. (1983) A bivariate spatial point pattern of ants' nests. Applied Statistics 32, 293–303

Lotwick, H. W. and Silverman, B. W. (1982). Methods for analysing spatial processes of several types of points. J. Royal Statist. Soc. Ser. B 44, 406–413.

Ripley, B.D. Statistical inference for spatial processes. Cambridge University Press, 1988.

Stoyan, D, Kendall, W.S. and Mecke, J. Stochastic geometry and its applications. 2nd edition. Springer Verlag, 1995.

Van Lieshout, M.N.M. and Baddeley, A.J. (1999) Indices of dependence between types in multivariate point patterns. Scandinavian Journal of Statistics 26, 511–532.

See Also

Gcross, Gdot, Gest

Examples


    data(longleaf)
     # Longleaf Pine data: marks represent diameter
    
    Gm <- Gmulti(longleaf, longleaf$marks <= 15, longleaf$marks >= 25)
    plot(Gm)

[Package spatstat version 1.10-3 Index]