Jmulti {spatstat}	R Documentation

Marked J Function

Description

For a marked point pattern, estimate the multitype J function summarising dependence between the points in subset I and those in subset J.

Usage


  Jmulti(X, I, J, eps=NULL, r=NULL, breaks=NULL, ..., disjoint=NULL)

Arguments

X 	 The observed point pattern, from which an estimate of the multitype distance distribution function JIJ(r) will be computed. It must be a marked point pattern. See under Details.

I 	 Subset of points of X from which distances are measured.

J 	 Subset of points in X to which distances are measured.

eps 	 A positive number. The pixel resolution of the discrete approximation to Euclidean distance (see Jest). There is a sensible default.

r 	 numeric vector. The values of the argument r at which the distribution function JIJ(r) should be evaluated. There is a sensible default. First-time users are strongly advised not to specify this argument. See below for important conditions on r.

breaks 	 An alternative to the argument r. Not normally invoked by the user. See the Details section.

... 	 Ignored.

disjoint 	 Optional flag indicating whether the subsets I and J are disjoint. If missing, this value will be computed by inspecting the vectors I and J

.

Details

The function Jmulti generalises Jest (for unmarked point patterns) and Jdot and Jcross (for multitype point patterns) to arbitrary marked point patterns.

Suppose X[I], X[J] are subsets, possibly overlapping, of a marked point process. Define

JIJ(r) = (1 - GIJ(r))/(1 - FJ(r))

where FJ(r) is the cumulative distribution function of the distance from a fixed location to the nearest point of X[J], and GJ(r) is the distribution function of the distance from a typical point of X[I] to the nearest distinct point of X[J].

The argument X must be a point pattern (object of class "ppp") or any data that are acceptable to as.ppp.

The arguments I and J specify two subsets of the point pattern. They may be logical vectors of length equal to X$n, or integer vectors with entries in the range 1 to X$n, etc.

It is assumed that X can be treated as a realisation of a stationary (spatially homogeneous) random spatial point process in the plane, observed through a bounded window. The window (which is specified in X as X$window) may have arbitrary shape. Biases due to edge effects are treated in the same manner as in Jest.

The argument r is the vector of values for the distance r at which JIJ(r) should be evaluated. It is also used to determine the breakpoints (in the sense of hist) for the computation of histograms of distances. The reduced-sample and Kaplan-Meier estimators are computed from histogram counts. In the case of the Kaplan-Meier estimator this introduces a discretisation error which is controlled by the fineness of the breakpoints.

First-time users would be strongly advised not to specify r. However, if it is specified, r must satisfy r[1] = 0, and max(r) must be larger than the radius of the largest disc contained in the window. Furthermore, the successive entries of r must be finely spaced.

Value

An object of class "fv" (see fv.object).
Essentially a data frame containing six numeric columns

r 	 the values of the argument r at which the function JIJ(r) has been estimated

rs 	 the ``reduced sample'' or ``border correction'' estimator of JIJ(r)

km 	 the spatial Kaplan-Meier estimator of JIJ(r)

un 	 the uncorrected estimate of JIJ(r), formed by taking the ratio of uncorrected empirical estimators of 1 - GIJ(r) and 1 - FJ(r), see Gdot and Fest.

theo 	 the theoretical value of JIJ(r) for a marked Poisson process with the same estimated intensity, namely 1.

Author(s)

Adrian Baddeley adrian@maths.uwa.edu.au http://www.maths.uwa.edu.au/~adrian/ and Rolf Turner rolf@math.unb.ca http://www.math.unb.ca/~rolf

References

Van Lieshout, M.N.M. and Baddeley, A.J. (1999) Indices of dependence between types in multivariate point patterns. Scandinavian Journal of Statistics 26, 511–532.

See Also

Jcross, Jdot, Jest

Examples


    data(longleaf)
     # Longleaf Pine data: marks represent diameter
    
    Jm <- Jmulti(longleaf, longleaf$marks <= 15, longleaf$marks >= 25)
    plot(Jm)

[Package spatstat version 1.10-3 Index]