Kdot {spatstat}	R Documentation

Multitype K Function (i-to-any)

Description

For a multitype point pattern, estimate the multitype K function which counts the expected number of other points of the process within a given distance of a point of type i.

Usage


Kdot(X, i=1, r=NULL, breaks=NULL, correction, ...)

Arguments

X 	 The observed point pattern, from which an estimate of the multitype K function Ki.(r) will be computed. It must be a multitype point pattern (a marked point pattern whose marks are a factor). See under Details.

i 	 Number or character string identifying the type (mark value) of the points in X from which distances are measured.

r 	 numeric vector. The values of the argument r at which the distribution function Ki.(r) should be evaluated. There is a sensible default. First-time users are strongly advised not to specify this argument. See below for important conditions on r.

breaks 	 An alternative to the argument r. Not normally invoked by the user. See the Details section.

correction 	 A character vector containing any selection of the options "border", "bord.modif", "isotropic", "Ripley" or "translate". It specifies the edge correction(s) to be applied.

... 	 Ignored.

Details

This function Kdot and its companions Kcross and Kmulti are generalisations of the function Kest to multitype point patterns.

A multitype point pattern is a spatial pattern of points classified into a finite number of possible ``colours'' or ``types''. In the spatstat package, a multitype pattern is represented as a single point pattern object in which the points carry marks, and the mark value attached to each point determines the type of that point.

The argument X must be a point pattern (object of class "ppp") or any data that are acceptable to as.ppp. It must be a marked point pattern, and the mark vector X$marks must be a factor. The argument i will be interpreted as a level of the factor X$marks. (Warning: this means that an integer value i=3 will be interpreted as the 3rd smallest level, not the number 3).

The ``type i to any type'' multitype K function of a stationary multitype point process X is defined so that lambda Ki.(r) equals the expected number of additional random points within a distance r of a typical point of type i in the process X. Here lambda is the intensity of the process, i.e. the expected number of points of X per unit area. The function Ki. is determined by the second order moment properties of X.

An estimate of Ki.(r) is a useful summary statistic in exploratory data analysis of a multitype point pattern. If the subprocess of type i points were independent of the subprocess of points of all types not equal to i, then Ki.(r) would equal pi * r^2. Deviations between the empirical Ki. curve and the theoretical curve pi * r^2 may suggest dependence between types.

This algorithm estimates the distribution function Ki.(r) from the point pattern X. It assumes that X can be treated as a realisation of a stationary (spatially homogeneous) random spatial point process in the plane, observed through a bounded window. The window (which is specified in X as X$window) may have arbitrary shape. Biases due to edge effects are treated in the same manner as in Kest, using the border correction.

The argument r is the vector of values for the distance r at which Ki.(r) should be evaluated. The values of r must be increasing nonnegative numbers and the maximum r value must exceed the radius of the largest disc contained in the window.

The pair correlation function can also be applied to the result of Kdot; see pcf.

Value

An object of class "fv" (see fv.object).
Essentially a data frame containing numeric columns

r 	 the values of the argument r at which the function Ki.(r) has been estimated

theo 	 the theoretical value of Ki.(r) for a marked Poisson process, namely pi * r^2

together with a column or columns named "border", "bord.modif", "iso" and/or "trans", according to the selected edge corrections. These columns contain estimates of the function Ki.(r) obtained by the edge corrections named.

Warnings

The argument i is interpreted as a level of the factor X$marks. Beware of the usual trap with factors: numerical values are not interpreted in the same way as character values. See the first example.

The reduced sample estimator of Ki. is pointwise approximately unbiased, but need not be a valid distribution function; it may not be a nondecreasing function of r. Its range is always within [0,1].

Author(s)

Adrian Baddeley adrian@maths.uwa.edu.au http://www.maths.uwa.edu.au/~adrian/ and Rolf Turner rolf@math.unb.ca http://www.math.unb.ca/~rolf

References

Cressie, N.A.C. Statistics for spatial data. John Wiley and Sons, 1991.

Diggle, P.J. Statistical analysis of spatial point patterns. Academic Press, 1983.

Harkness, R.D and Isham, V. (1983) A bivariate spatial point pattern of ants' nests. Applied Statistics 32, 293–303

Lotwick, H. W. and Silverman, B. W. (1982). Methods for analysing spatial processes of several types of points. J. Royal Statist. Soc. Ser. B 44, 406–413.

Ripley, B.D. Statistical inference for spatial processes. Cambridge University Press, 1988.

Stoyan, D, Kendall, W.S. and Mecke, J. Stochastic geometry and its applications. 2nd edition. Springer Verlag, 1995.

See Also

Kdot, Kest, Kmulti, pcf

Examples


     # Lansing woods data: 6 types of trees
    data(lansing)

    ## Not run: 
    Kh. <- Kdot(lansing, "hickory") 
    
## End(Not run)
    
    # diagnostic plot for independence between hickories and other trees
    plot(Kh.)

    # synthetic example with two marks "a" and "b"
    pp <- runifpoispp(50)
    pp <- pp %mark% sample(c("a","b"), pp$n, replace=TRUE)
    K <- Kdot(pp, "a")

[Package spatstat version 1.10-3 Index]