MultiStraussHard {spatstat}	R Documentation

The Multitype/Hard Core Strauss Point Process Model

Description

Creates an instance of the multitype/hard core Strauss point process model which can then be fitted to point pattern data.

Usage


  MultiStraussHard(types, iradii, hradii)

Arguments

types 	 Vector of all possible types (i.e. the possible levels of the marks variable in the data)

iradii 	 Matrix of interaction radii

hradii 	 Matrix of hard core radii

Details

This is a hybrid of the multitype Strauss process (see MultiStrauss) and the hard core process (case gamma = 0 of the Strauss process). A pair of points of types i and j must not lie closer than h[i,j] units apart; if the pair lies more than h[i,j] and less than r[i,j] units apart, it contributes a factor gamma[i,j] to the probability density.

The matrices iradii and hradii must be symmetric, with entries which are either positive numbers or NA. A value of NA indicates that no interaction term should be included for this combination of types.

Note that only the interaction radii and hardcore radii are specified in MultiStraussHard. The canonical parameters log(beta[j]) and log(gamma[i,j]) are estimated by ppm(), not fixed in MultiStraussHard().

Value

An object of class "interact" describing the interpoint interaction structure of the multitype/hard core Strauss process with interaction radii iradii[i,j] and hard core radii hradii[i,j].

Warnings

The argument types is interpreted as a set of factor levels. That is, in order that ppm can fit the multitype Strauss model correctly to a point pattern X, this must be a marked point pattern; the mark vector X$marks must be a factor; and the argument types must equal levels(X$marks).

Author(s)

Adrian Baddeley adrian@maths.uwa.edu.au http://www.maths.uwa.edu.au/~adrian/ and Rolf Turner rolf@math.unb.ca http://www.math.unb.ca/~rolf

See Also

ppm, pairwise.family, ppm.object, MultiStrauss, Strauss

Examples


   r <- matrix(3, nrow=2,ncol=2)
   h <- matrix(c(1,2,2,1), nrow=2,ncol=2)
   MultiStraussHard(1:2, r, h)
   # prints a sensible description of itself
   data(betacells)
   r <- 30.0 * matrix(c(1,2,2,1), nrow=2,ncol=2)
   h <- 30.0 * matrix(c(NA,1,1,NA), nrow=2,ncol=2)
   ppm(betacells, ~1, MultiStraussHard(c("off","on"), r, h), rbord=60.0)
   # fit the stationary multitype hardcore Strauss process to `betacells'

[Package spatstat version 1.10-3 Index]