OrdThresh {spatstat}	R Documentation

Ord's Interaction model

Description

Creates an instance of Ord's point process model which can then be fitted to point pattern data.

Usage


  OrdThresh(r)

Arguments

r 	 Positive number giving the threshold value for Ord's model.

Details

Ord's point process model (Ord, 1977) is a Gibbs point process of infinite order. Each point x[i] in the point pattern x contributes a factor g(a[i]) where a[i] = a(x[i], x) is the area of the tile associated with x[i] in the Dirichlet tessellation of x. The function g is simply g(a) = 1 if a >= r and g(a) = gamma < 1 if a < r, where r is called the threshold value.

This function creates an instance of Ord's model with a given value of r. It can then be fitted to point process data using ppm.

Value

An object of class "interact" describing the interpoint interaction structure of a point process.

Author(s)

Adrian Baddeley adrian@maths.uwa.edu.au http://www.maths.uwa.edu.au/~adrian/ and Rolf Turner rolf@math.unb.ca http://www.math.unb.ca/~rolf

References

Baddeley, A. and Turner, R. (2000) Practical maximum pseudolikelihood for spatial point patterns. Australian and New Zealand Journal of Statistics 42, 283–322.

Ord, J.K. (1977) Contribution to the discussion of Ripley (1977).

Ord, J.K. (1978) How many trees in a forest? Mathematical Scientist 3, 23–33.

Ripley, B.D. (1977) Modelling spatial patterns (with discussion). Journal of the Royal Statistical Society, Series B, 39, 172 – 212.

See Also

ppm, ppm.object

[Package spatstat version 1.10-3 Index]