PairPiece {spatstat}	R Documentation

The Piecewise Constant Pairwise Interaction Point Process Model

Description

Creates an instance of a pairwise interaction point process model with piecewise constant potential function. The model can then be fitted to point pattern data.

Usage


  PairPiece(r)

Arguments

r 	 vector of jump points for the potential function

Details

A pairwise interaction point process in a bounded region is a stochastic point process with probability density of the form

f(x_1,...,x_n) = alpha . product { b(x[i]) } product { h(x_i, x_j) }

where x[1],...,x[n] represent the points of the pattern. The first product on the right hand side is over all points of the pattern; the second product is over all unordered pairs of points of the pattern.

Thus each point x[i] of the pattern contributes a factor b(x[i]) to the probability density, and each pair of points x[i], x[j] contributes a factor h(x[i], x[j]) to the density.

The pairwise interaction term h(u, v) is called piecewise constant if it depends only on the distance between u and v, say h(u,v) = H(||u-v||), and H is a piecewise constant function (a function which is constant except for jumps at a finite number of places).

The function ppm(), which fits point process models to point pattern data, requires an argument of class "interact" describing the interpoint interaction structure of the model to be fitted. The appropriate description of the piecewise constant pairwise interaction is yielded by the function PairPiece(). See the examples below.

The entries of r must be strictly increasing, positive numbers. They are interpreted as the points of discontinuity of H. It is assumed that H(s) =1 for all s > rmax where rmax is the maximum value in r. Thus the model has as many regular parameters (see ppm) as there are entries in r. The i-th regular parameter theta[i] is the logarithm of the value of the interaction function H on the interval [r[i-1],r[i]).

If r is a single number, this model is similar to the Strauss process, see Strauss. The difference is that in PairPiece the interaction function is continuous on the right, while in Strauss it is continuous on the left.

The analogue of this model for multitype point processes has not yet been implemented.

Value

An object of class "interact" describing the interpoint interaction structure of a point process. The process is a pairwise interaction process, whose interaction potential is piecewise constant, with jumps at the distances given in the vector r.

Author(s)

Adrian Baddeley adrian@maths.uwa.edu.au http://www.maths.uwa.edu.au/~adrian/ and Rolf Turner rolf@math.unb.ca http://www.math.unb.ca/~rolf

See Also

ppm, pairwise.family, ppm.object, Strauss rmh.ppm

Examples


   PairPiece(c(0.1,0.2))
   # prints a sensible description of itself
   data(cells) 
   ppm(cells, ~1, PairPiece(r = c(0.05, 0.1, 0.2)), rbord=0.2)
   # fit a stationary piecewise constant pairwise interaction process
   ppm(cells, ~polynom(x,y,3), PairPiece(c(0.05, 0.1)), rbord=0.1)
   # nonstationary process with log-cubic polynomial trend

[Package spatstat version 1.10-3 Index]