Poisson {spatstat}	R Documentation

Poisson Point Process Model

Description

Creates an instance of the Poisson point process model which can then be fitted to point pattern data.

Usage


 Poisson()

Details

The function ppm(), which fits point process models to point pattern data, requires an argument interaction of class "interact" describing the interpoint interaction structure of the model to be fitted. The appropriate description of the Poisson process is provided by the value of the function Poisson().

This works for all types of Poisson processes including multitype and nonstationary Poisson processes.

Value

An object of class "interact" describing the interpoint interaction structure of the Poisson point process (namely, there are no interactions).

Author(s)

Adrian Baddeley adrian@maths.uwa.edu.au http://www.maths.uwa.edu.au/~adrian/ and Rolf Turner rolf@math.unb.ca http://www.math.unb.ca/~rolf

See Also

Strauss, StraussHard

Examples


 data(nztrees)
 ppm(nztrees, ~1, Poisson())
 # fit the stationary Poisson process to 'nztrees'
 # no edge correction needed

 data(longleaf)
 
 longadult <- longleaf[longleaf$marks >= 30, ]
 longadult <- unmark(longadult)
 ppm(longadult, ~ x, Poisson())
 # fit the nonstationary Poisson process 
 # with intensity lambda(x,y) = exp( a + bx)

 data(lansing)
 # trees marked by species
 
 ppm(lansing, ~ marks, Poisson())
 # fit stationary marked Poisson process
 # with different intensity for each species

## Not run: 
 ppm(lansing, ~ marks * polynom(x,y,3), Poisson())
## End(Not run)
 # fit nonstationary marked Poisson process
 # with different log-cubic trend for each species


[Package spatstat version 1.10-3 Index]