Softcore {spatstat}	R Documentation

The Soft Core Point Process Model

Description

Creates an instance of the Soft Core point process model which can then be fitted to point pattern data.

Usage


  Softcore(kappa)

Arguments

kappa 	 The exponent kappa of the Soft Core interaction

Details

The (stationary) Soft Core point process with parameters beta and sigma and exponent kappa is the pairwise interaction point process in which each point contributes a factor beta to the probability density of the point pattern, and each pair of points contributes a factor

exp( - (sigma/d)^(2/kappa) )

to the density, where d is the distance between the two points.

Thus the process has probability density

f(x_1,...,x_n) = alpha . beta^n(x) exp( - sum (sigma/||x[i]-x[j]||)^(2/kappa))

where x[1],...,x[n] represent the points of the pattern, n(x) is the number of points in the pattern, alpha is the normalising constant, and the sum on the right hand side is over all unordered pairs of points of the pattern.

This model describes an ``ordered'' or ``inhibitive'' process, with the interpoint interaction decreasing smoothly with distance. The strength of interaction is controlled by the parameter sigma, a positive real number, with larger values corresponding to stronger interaction; and by the exponent kappa in the range (0,1), with larger values corresponding to weaker interaction. If sigma = 0 the model reduces to the Poisson point process. If sigma > 0, the process is well-defined only for kappa in (0,1). The limit of the model as kappa -> 0 is the hard core process with hard core distance h=sigma.

The nonstationary Soft Core process is similar except that the contribution of each individual point x[i] is a function beta(x[i]) of location, rather than a constant beta.

The function ppm(), which fits point process models to point pattern data, requires an argument of class "interact" describing the interpoint interaction structure of the model to be fitted. The appropriate description of the Soft Core process pairwise interaction is yielded by the function Softcore(). See the examples below.

Note the only argument is the exponent kappa. When kappa is fixed, the model becomes an exponential family with canonical parameters log(beta) and

log(gamma) = (2/kappa) log(sigma)

The canonical parameters are estimated by ppm(), not fixed in Softcore().

Value

An object of class "interact" describing the interpoint interaction structure of the Soft Core process with exponent kappa.

Author(s)

Adrian Baddeley adrian@maths.uwa.edu.au http://www.maths.uwa.edu.au/~adrian/ and Rolf Turner rolf@math.unb.ca http://www.math.unb.ca/~rolf

References

Ogata, Y, and Tanemura, M. (1981). Estimation of interaction potentials of spatial point patterns through the maximum likelihood procedure. Annals of the Institute of Statistical Mathematics, B 33, 315–338.

Ogata, Y, and Tanemura, M. (1984). Likelihood analysis of spatial point patterns. Journal of the Royal Statistical Society, series B 46, 496–518.

See Also

ppm, pairwise.family, ppm.object

Examples


   data(cells) 
   ppm(cells, ~1, Softcore(kappa=0.5), rbord=0)
   # fit the stationary Soft Core process to `cells'

[Package spatstat version 1.10-3 Index]