Strauss {spatstat}	R Documentation

The Strauss Point Process Model

Description

Creates an instance of the Strauss point process model which can then be fitted to point pattern data.

Usage


  Strauss(r)

Arguments

r 	 The interaction radius of the Strauss process

Details

The (stationary) Strauss process with interaction radius r and parameters beta and gamma is the pairwise interaction point process in which each point contributes a factor beta to the probability density of the point pattern, and each pair of points closer than r units apart contributes a factor gamma to the density.

Thus the probability density is

f(x_1,...,x_n) = alpha . beta^n(x) gamma^s(x)

where x[1],...,x[n] represent the points of the pattern, n(x) is the number of points in the pattern, s(x) is the number of distinct unordered pairs of points that are closer than r units apart, and alpha is the normalising constant.

The interaction parameter gamma must be less than or equal to 1 so that this model describes an ``ordered'' or ``inhibitive'' pattern.

The nonstationary Strauss process is similar except that the contribution of each individual point x[i] is a function beta(x[i]) of location, rather than a constant beta.

The function ppm(), which fits point process models to point pattern data, requires an argument of class "interact" describing the interpoint interaction structure of the model to be fitted. The appropriate description of the Strauss process pairwise interaction is yielded by the function Strauss(). See the examples below.

Note the only argument is the interaction radius r. When r is fixed, the model becomes an exponential family. The canonical parameters log(beta) and log(gamma) are estimated by ppm(), not fixed in Strauss().

Value

An object of class "interact" describing the interpoint interaction structure of the Strauss process with interaction radius r.

Author(s)

Adrian Baddeley adrian@maths.uwa.edu.au http://www.maths.uwa.edu.au/~adrian/ and Rolf Turner rolf@math.unb.ca http://www.math.unb.ca/~rolf

See Also

ppm, pairwise.family, ppm.object

Examples


   Strauss(r=0.1)
   # prints a sensible description of itself
   data(cells) 
   ppm(cells, ~1, Strauss(r=0.07), rbord=0.07)
   # fit the stationary Strauss process to `cells'
   ppm(cells, ~polynom(x,y,3), Strauss(r=0.07), rbord=0.1)
   # fit a nonstationary Strauss process with log-cubic polynomial trend

[Package spatstat version 1.10-3 Index]