StraussHard {spatstat}	R Documentation

The Strauss / Hard Core Point Process Model

Description

Creates an instance of the ``Strauss/ hard core'' point process model which can then be fitted to point pattern data.

Usage


  StraussHard(r, hc)

Arguments

r 	 The interaction radius of the Strauss interaction

hc 	 The hard core distance

Details

A Strauss/hard core process with interaction radius r, hard core distance h < r, and parameters beta and gamma, is a pairwise interaction point process in which

distinct points are not allowed to come closer than a distance h apart
each pair of points closer than r units apart contributes a factor gamma to the probability density.

This is a hybrid of the Strauss process and the hard core process.

The probability density is zero if any pair of points is closer than h units apart, and otherwise equals

f(x_1,...,x_n) = alpha . beta^n(x) gamma^s(x)

where x[1],...,x[n] represent the points of the pattern, n(x) is the number of points in the pattern, s(x) is the number of distinct unordered pairs of points that are closer than r units apart, and alpha is the normalising constant.

The interaction parameter gamma may take any positive value (unlike the case for the Strauss process). If gamma = 1, the process reduces to a classical hard core process. If gamma < 1, the model describes an ``ordered'' or ``inhibitive'' pattern. If gamma > 1, the model is ``ordered'' or ``inhibitive'' up to the distance h, but has an ``attraction'' between points lying at distances in the range between h and r.

The function ppm(), which fits point process models to point pattern data, requires an argument of class "interact" describing the interpoint interaction structure of the model to be fitted. The appropriate description of the Strauss/hard core process pairwise interaction is yielded by the function StraussHard(). See the examples below.

The canonical parameter log(gamma) is estimated by ppm(), not fixed in StraussHard().

Value

An object of class "interact" describing the interpoint interaction structure of the ``Strauss/hard core'' process with Strauss interaction radius r and hard core distance hc.

Author(s)

Adrian Baddeley adrian@maths.uwa.edu.au http://www.maths.uwa.edu.au/~adrian/ and Rolf Turner rolf@math.unb.ca http://www.math.unb.ca/~rolf

References

Baddeley, A. and Turner, R. (2000) Practical maximum pseudolikelihood for spatial point patterns. Australian and New Zealand Journal of Statistics 42, 283–322.

Ripley, B.D. (1981) Spatial statistics. John Wiley and Sons.

Strauss, D.J. (1975) A model for clustering. Biometrika 63, 467–475.

See Also

ppm, pairwise.family, ppm.object

Examples


   StraussHard(r=1,hc=0.02)
   # prints a sensible description of itself

   data(cells) 
   ppm(cells, ~1, StraussHard(r=0.1, hc=0.05), rbord=0.1)
   # fit the stationary Strauss/hard core  process to `cells'

   ppm(cells, ~ polynom(x,y,3), StraussHard(r=0.1, hc=0.05), rbord=0.1)
   # fit a nonstationary Strauss/hard core process
   # with log-cubic polynomial trend
   

[Package spatstat version 1.10-3 Index]