affine.owin {spatstat}	R Documentation

Apply Affine Transformation To Window

Description

Applies any affine transformation of the plane (linear transformation plus vector shift) to a window.

Usage


  ## S3 method for class 'owin':
  affine(X, mat=diag(c(1,1)), vec=c(0,0), ...)

Arguments

X 	 Window (object of class "owin").

mat 	 Matrix representing a linear transformation.

vec 	 Vector of length 2 representing a translation.

... 	 Ignored

Details

The window is subjected first to the linear transformation represented by mat (multiplying on the left by mat), and then the result is translated by the vector vec.

The argument mat must be a nonsingular 2 * 2 matrix.

This is a method for the generic function affine.

Value

Another window (of class "owin") representing the result of applying the affine transformation.

Author(s)

Adrian Baddeley adrian@maths.uwa.edu.au http://www.maths.uwa.edu.au/~adrian/ and Rolf Turner rolf@math.unb.ca http://www.math.unb.ca/~rolf

See Also

affine, affine.ppp, rotate, shift

Examples


  # shear transformation
  X <- affine(owin(), matrix(c(1,0,0.6,1),ncol=2))
  ## Not run: 
  plot(X)
  
## End(Not run)

[Package spatstat version 1.10-3 Index]