allstats {spatstat}	R Documentation

Calculate four standard summary functions of a point pattern.

Description

Calculates the F, G, J, and K summary functions for an unmarked point pattern. Returns them as a function array (of class "fasp", see fasp.object).

Usage


  allstats(pp, ..., dataname=NULL, verb=FALSE)

Arguments

pp 	 The observed point pattern, for which summary function estimates are required. An object of class "ppp". It must not be marked.

... 	 Optional arguments passed to the summary functions Fest, Gest, Jest and Kest.

dataname 	 A character string giving an optional (alternative) name for the point pattern.

verb 	 A logical value meaning ``verbose''. If TRUE, progress reports are printed during calculation.

Details

This computes four standard summary statistics for a point pattern: the empty space function F(r), nearest neighbour distance distribution function G(r), van Lieshout-Baddeley function J(r) and Ripley's function K(r). The real work is done by Fest, Gest, Jest and Kest respectively. Consult the help files for these functions for further information about the statistical interpretation of F, G, J and K.

If verb is TRUE, then ``progress reports'' (just indications of completion) are printed out when the calculations are finished for each of the four function types.

The overall title of the array of four functions (for plotting by plot.fasp) will be formed from the argument dataname. If this is not given, it defaults to the expression for pp given in the call to allstats.

Value

A function array (an object of class "fasp", see fasp.object). This can be plotted using plot.fasp.
The function array has dimensions 2 * 2 with the entries at positions [1,1], [1,2], [2,1] and [2,2] representing F(r), G(r), J(r) and K(r) respectively.
Each function entry fns[[i]] retains the format of the output of the relevant estimating routine Fest, Gest, Jest or Kest.
The default formulae for plotting these functions are cbind(km,theo) ~ r for F, G, and J, and cbind(trans,theo) ~ r for K.

Author(s)

Adrian Baddeley adrian@maths.uwa.edu.au http://www.maths.uwa.edu.au/~adrian/ and Rolf Turner rolf@math.unb.ca http://www.math.unb.ca/~rolf

See Also

plot.fasp, fasp.object, Fest, Gest, Jest, Kest, alltypes

Examples


        data(swedishpines)
        a <- allstats(swedishpines,dataname="Swedish Pines")
        ## Not run: 
        plot(a)
        plot(a, subset=list("r<=15","r<=15","r<=15","r<=50"))
        
## End(Not run)

[Package spatstat version 1.10-3 Index]