angles.psp {spatstat}	R Documentation

Orientation Angles of Line Segments

Description

Computes the orientation angle of each line segment in a line segment pattern.

Usage


  angles.psp(x, directed=FALSE)

Arguments

x 	 A line segment pattern (object of class "psp").

directed 	 Logical flag. See details.

Details

For each line segment, the angle of inclination to the x-axis (in radians) is computed, and the angles are returned as a numeric vector.

If directed=TRUE, the directed angle of orientation is computed. The angle respects the sense of direction from (x0,y0) to (x1,y1). The values returned are angles in the full range from -pi to pi. The angle is computed as atan2(y1-y0,x1-x0). See atan2.

If directed=FALSE, the undirected angle of orientation is computed. Angles differing by pi are regarded as equivalent. The values returned are angles in the range from 0 to pi. These angles are computed by first computing the directed angle, then adding pi to any negative angles.

Value

Numeric vector.

Author(s)

Adrian Baddeley adrian@maths.uwa.edu.au http://www.maths.uwa.edu.au/~adrian/ and Rolf Turner rolf@math.unb.ca http://www.math.unb.ca/~rolf

See Also

summary.psp, midpoints.psp, lengths.psp

Examples


  a <- psp(runif(10), runif(10), runif(10), runif(10), window=owin())
  b <- angles.psp(a)   

[Package spatstat version 1.10-3 Index]