anova.ppm {spatstat}	R Documentation

ANOVA for Fitted Point Process Models

Description

Performs analysis of deviance for two or more fitted point process models.

Usage


  ## S3 method for class 'ppm':
  anova(object, ..., test=NULL, override=FALSE)

Arguments

object 	 A fitted point process model (object of class "ppm").

... 	 One or more fitted point process models.

test 	 Character string, partially matching one of "Chisq", "F" or "Cp".

override 	 Logical flag indicating whether to proceed even when there is no statistical theory to support the calculation.

Details

This is a method for anova for fitted point process models (objects of class "ppm", usually generated by the model-fitting function ppm).

If the fitted models are all Poisson point processes, then this function performs an Analysis of Deviance of the fitted models. The output shows the deviance differences (i.e. 2 times log likelihood ratio), the difference in degrees of freedom, and (if test="Chi") the two-sided p-values for the chi-squared tests. Their interpretation is very similar to that in anova.glm.

If some of the fitted models are not Poisson point processes, then there is no statistical theory available to support a similar analysis. The function issues a warning, and (by default) returns a NULL value.

However if override=TRUE, then a kind of analysis of deviance table will be printed. The `deviance' differences in this table are equal to 2 times the differences in the maximised values of the log pseudolikelihood (see ppm). At the time of writing, there is no statistical theory to support inferential interpretation of log pseudolikelihood ratios. The override option is provided for research purposes only!

Value

An object of class "anova", or NULL.

Author(s)

Adrian Baddeley adrian@maths.uwa.edu.au http://www.maths.uwa.edu.au/~adrian/ and Rolf Turner rolf@math.unb.ca http://www.math.unb.ca/~rolf

See Also

ppm

Examples


 data(swedishpines)
 mod0 <- ppm(swedishpines, ~1, Poisson())
 modx <- ppm(swedishpines, ~x, Poisson())
 anova.ppm(modx, mod0, test="Chi")

[Package spatstat version 1.10-3 Index]