append.psp {spatstat}	R Documentation

Combine Two Line Segment Patterns

Description

Combine two line segment patterns into a single pattern.

Usage


  append.psp(A, B)

Arguments

A,B 	 Line segment patterns (objects of class "psp").

Details

This function is used to superimpose two line segment patterns A and B.

The two patterns must have identical windows. If one pattern has marks, then the other must also have marks of the same type.

(To combine two point patterns, see superimpose).

Value

Another line segment pattern (object of class "psp").

Author(s)

Adrian Baddeley adrian@maths.uwa.edu.au http://www.maths.uwa.edu.au/~adrian/ and Rolf Turner rolf@math.unb.ca http://www.math.unb.ca/~rolf

See Also

psp, as.psp, superimpose,

Examples


  X <- psp(runif(20), runif(20), runif(20), runif(20),  window=owin())
  Y <- psp(runif(5), runif(5), runif(5), runif(5),  window=owin())
  append.psp(X,Y)

[Package spatstat version 1.10-3 Index]