area.owin {spatstat}	R Documentation

Area of a Window

Description

Computes the area of a window

Usage


 area.owin(w)

Arguments

w 	 A window, whose area will be computed. This should be an object of class owin, or can be given in any format acceptable to as.owin().

Details

If the window W is of type "rectangle" or "polygonal", the area of this rectangular window is computed by analytic geometry. If W is of type "mask" the area of the discrete raster approximation of the window is computed by summing the binary image values and adjusting for pixel size.

Value

The numerical value of the area of the window.

Author(s)

Adrian Baddeley adrian@maths.uwa.edu.au http://www.maths.uwa.edu.au/~adrian/ and Rolf Turner rolf@math.unb.ca http://www.math.unb.ca/~rolf

See Also

owin.object, as.owin

Examples


  w <- unit.square()
  area.owin(w)
       # returns 1.00000

  k <- 6
  theta <- 2 * pi * (0:(k-1))/k
  co <- cos(theta)
  si <- sin(theta)
  mas <- owin(c(-1,1), c(-1,1), poly=list(x=co, y=si))
  area.owin(mas)
      # returns approx area of k-gon
  
  mas <- as.mask(w, eps=0.01)
  X <- raster.x(mas)
  Y <- raster.y(mas)
  mas$m <- ((X - 0.5)^2 + (Y - 0.5)^2 <= 1)
  area.owin(mas)
       # returns 3.14 approx     


[Package spatstat version 1.10-3 Index]