as.owin {spatstat}	R Documentation

Convert Data To Class owin

Description

Converts data specifying an observation window in any of several formats, into an object of class "owin".

Usage


 as.owin(W, ..., fatal=TRUE)
 ## S3 method for class 'owin':
 as.owin(W, ..., fatal=TRUE)
 ## S3 method for class 'ppp':
 as.owin(W, ..., fatal=TRUE)
 ## S3 method for class 'im':
 as.owin(W, ..., fatal=TRUE)
 ## Default S3 method:
 as.owin(W, ..., fatal=TRUE)

Arguments

W 	 Data specifying an observation window, in any of several formats described under Details below.

fatal 	 Logical flag determining what to do if the data cannot be converted to an observation window. See Details.

... 	 Ignored.

Details

The class "owin" is a way of specifying the observation window for a point pattern. See owin.object for an overview.

This function converts data in any of several formats into an object of class "owin" for use by the spatstat package. The argument W may be

an object of class "owin"
a structure with entries xrange, yrange specifying the x and y dimensions of a rectangle
a four-element vector (interpreted as (xmin, xmax, ymin, ymax)) specifying the x and y dimensions of a rectangle
a structure with entries xl, xu, yl, yu specifying the x and y dimensions of a rectangle as (xmin, xmax) = (xl, xu) and (ymin, ymax) = (yl, yu). This will accept objects of class spp used in the Venables and Ripley spatial library.
an object of class "ppp" representing a point pattern. In this case, the object's window structure will be extracted.
an object of class "im" representing a pixel image. In this case, a window of type "mask" will be returned, with the same pixel raster coordinates as the image. An image pixel value of NA, signifying that the pixel lies outside the window, is transformed into the logical value FALSE, which is the corresponding convention for window masks.

If the argument W is not in one of these formats and cannot be converted to a window, then an error will be generated (if fatal=TRUE) or a value of NULL will be returned (if fatal=FALSE).

The function as.owin is generic, with methods for "owin", "im" and "ppp" as well as the default method.

Value

An object of class "owin" (see owin.object) specifying an observation window.

Author(s)

Adrian Baddeley adrian@maths.uwa.edu.au http://www.maths.uwa.edu.au/~adrian/ and Rolf Turner rolf@math.unb.ca http://www.math.unb.ca/~rolf

See Also

owin.object, owin

Examples


 w <- as.owin(c(0,1,0,1))
 w <- as.owin(list(xrange=c(0,5),yrange=c(0,10)))
 # point pattern
 data(demopat)
 w <- as.owin(demopat)
 # image
 Z <- as.im(function(x,y) { x + 3}, unit.square())
 w <- as.owin(Z)

 # Venables & Ripley 'spatial' package
 require(spatial)
 towns <- ppinit("towns.dat")
 w <- as.owin(towns)
 detach(package:spatial)

[Package spatstat version 1.10-3 Index]