bdist.points {spatstat}	R Documentation

Distance to Boundary of Window

Description

Computes the distances from each point of a point pattern to the boundary of the window.

Usage


 bdist.points(X)

Arguments

X 	 A point pattern (object of class "ppp").

Details

This function computes, for each point x[i] in the point pattern X, the shortest distance dist(x[i], W') from x[i] to the boundary of the window W of observation.

If the window X$window is of type "rectangle" or "polygonal", then these distances are computed by analytic geometry and are exact, up to rounding errors. If the window is of type "mask" then the distances are computed using the real-valued distance transform, which is an approximation with maximum error equal to the width of one pixel in the mask.

Value

A numeric vector, giving the distances from each point of the pattern to the boundary of the window.

Author(s)

Adrian Baddeley adrian@maths.uwa.edu.au http://www.maths.uwa.edu.au/~adrian/ and Rolf Turner rolf@math.unb.ca http://www.math.unb.ca/~rolf

See Also

ppp.object, erode.owin

Examples


  data(cells)
  d <- bdist.points(cells)

[Package spatstat version 1.10-3 Index]