beginner {spatstat}	R Documentation

Print Introduction For Beginners

Description

Prints an introduction for beginners to the spatstat package, or another specified package.

Usage

beginner(package = "spatstat")

Arguments

package
Name of package.

Details

This function prints an introduction for beginners to the spatstat package.

The function can be executed simply by typing beginner without parentheses.

If the argument package is given, then the function prints the beginner's help file BEGINNER.txt from the specified package (if it has one).

Value

Null.

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au

and Rolf Turner rolfturner@posteo.net

See Also

latest.news

Examples

  beginner

[Package spatstat version 3.3-1 Index]