bugfixes {spatstat}	R Documentation

List Recent Bug Fixes

Description

List all bug fixes in a package, starting from a certain date or version of the package. Fixes are sorted alphabetically by the name of the affected function. The default is to list bug fixes in the latest version of the spatstat family of packages.

Usage

  bugfixes(sinceversion = NULL, sincedate = NULL,
           package = spatstat.family(), show = TRUE)

Arguments

sinceversion
Earliest version of package for which bugs should be listed. A character string. The default is the current installed version.

sincedate
Earliest release date of package for which bugs should be listed. A character string or a date-time object.

package
The name of the package (or packages) for which bugs are to be listed. A character string or character vector.

show
Logical value indicating whether to display the bug table on the terminal.

Details

Bug reports are extracted from the NEWS file of the specified package. Only those after a specified date, or after a specified version of the package, are retained. The bug reports are then sorted alphabetically, so that all bugs affecting a particular function are listed consecutively. Finally the table of bug reports is displayed (if show=TRUE) and returned invisibly.

The argument sinceversion should be a character string like "1.2-3". The default is the current installed version of the package.

The argument sincedate should be a character string like "2015-05-27", or a date-time object.

If sinceversion="all" or sincedate="all" then all recorded bugs will be listed.

The special options sinceversion="book" and sincedate="book" are interpreted to mean sincedate="2015-06-05", which gives all bugs reported after publication of the book by Baddeley, Rubak and Turner (2015).

Typing bugfixes without parentheses will display a table of all bugs that were fixed in the current installed version of spatstat and its sub-packages.

By default, bugs in the extension packages spatstat.local, spatstat.Knet, spatstat.gui are not reported. To include these bugs as well, set package=spatstat.family(TRUE,TRUE).

Value

(Invisibly) a data frame, belonging to the class "bugtable", which has a print method.

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au.

References

Baddeley, A., Rubak, E. and Turner, R. (2015) Spatial Point Patterns: Methodology and Applications with R. Chapman and Hall/CRC Press.

See Also

latest.changes, latest.news, news.

Examples

   bugfixes
   ## show all bugs reported after publication of the spatstat book
   if(interactive()) bugfixes(sinceversion="book")

[Package spatstat version 3.3-1 Index]