cells {spatstat}	R Documentation

Biological Cells Point Pattern

Description

The data record the locations of the centres of 42 biological cells observed under optical microscopy in a histological section. The microscope field-of-view has been rescaled to the unit square.

The data were recorded by F.H.C. Crick and B.D. Ripley, and analysed in Ripley (1977, 1981) and Diggle (1983). They are often used as a canonical example of an `ordered' point pattern.

Usage

data(cells)

Format

An object of class "ppp" representing the point pattern of cell centres. See ppp.object for details of the format.

Source

Crick and Ripley, see Ripley (1977)

References

Diggle, P.J. (1983) Statistical analysis of spatial point patterns. Academic Press.

Ripley, B.D. (1977) Modelling spatial patterns (with discussion). Journal of the Royal Statistical Society, Series B 39, 172–212.

Ripley, B.D. (1981) Spatial statistics. John Wiley and Sons.

[Package spatstat version 1.10-3 Index]