coef.ppm {spatstat}	R Documentation

Coefficients of Fitted Point Process Model

Description

Given a point process model fitted to a point pattern, extract the coefficients of the fitted model. A method for coef.

Usage


  ## S3 method for class 'ppm':
  coef(object, ...)

Arguments

object 	 The fitted point process model (an object of class "ppm")

... 	 Ignored.

Details

This function is a method for the generic function coef.

The argument object must be a fitted point process model (object of class "ppm"). Such objects are produced by the maximum pseudolikelihood fitting algorithm ppm).

This function extracts the vector of coefficients of the fitted model. This is the estimate of the parameter vector theta such that the conditional intensity of the model is of the form

lambda(u,x) = exp(theta . S(u,x))

where S(u,x) is a (vector-valued) statistic.

For example, if the model object is the uniform Poisson process, then coef(object) will yield a single value (named "(Intercept)") which is the logarithm of the fitted intensity of the Poisson process.

Use print.ppm to print a more useful description of the fitted model.

Value

A vector containing the fitted coefficients.

Author(s)

Adrian Baddeley adrian@maths.uwa.edu.au http://www.maths.uwa.edu.au/~adrian/ and Rolf Turner rolf@math.unb.ca http://www.math.unb.ca/~rolf

See Also

print.ppm, ppm.object, ppm

Examples


    data(cells)

    poi <- ppm(cells, ~1, Poisson())
    coef(poi)
    # This is the log of the fitted intensity of the Poisson process

    stra <- ppm(cells, ~1, Strauss(r=0.07), rbord=0.07)
    coef(stra)

    # The two entries "(Intercept)" and "Interaction"
    # are respectively log(beta) and log(gamma)
    # in the usual notation for Strauss(beta, gamma, r)


[Package spatstat version 1.10-3 Index]