compatible.fv {spatstat}	R Documentation

Test Whether Two Function Objects Are Compatible

Description

Tests whether two function objects (class "fv") are compatible.

Usage


  compatible.fv(A, B)

Arguments

A,B 	 Function value objects (class "fv").

Details

An object of class "fv" is essentially a data frame containing several different statistical estimates of the same function. Such objects are returned by Kest and its relatives.

This function tests whether two such objects are compatible (so that, for example, they could be added or subtracted).

The functions are compatible if they have been evaluated at the same sequence of values of the argument r, and if the statistical estimates have the same names.

Value

Logical value: TRUE if the objects are compatible, and FALSE if they are not.

Author(s)

Adrian Baddeley adrian@maths.uwa.edu.au http://www.maths.uwa.edu.au/~adrian/ and Rolf Turner rolf@math.unb.ca http://www.math.unb.ca/~rolf

[Package spatstat version 1.10-3 Index]