compatible.im {spatstat}	R Documentation

Test Whether Two Pixel Images Are Compatible

Description

Tests whether two pixel image objects have compatible dimensions.

Usage


compatible.im(A, B, tol=1e-6)

Arguments

A,B 	 Pixel images (objects of class "im").

tol 	 Tolerance factor

Details

This function tests whether the pixel images A and B have compatible pixel dimensions. They are compatible if they have the same number of rows and columns, the same physical pixel dimensions, and occupy the same rectangle in the plane.

The argument tol specifies the maximum tolerated error in the pixel coordinates, expressed as a fraction of the dimensions of a single pixel.

Value

Logical value: TRUE if the images are compatible, and FALSE if they are not.

Author(s)

Adrian Baddeley adrian@maths.uwa.edu.au http://www.maths.uwa.edu.au/~adrian/ and Rolf Turner rolf@math.unb.ca http://www.math.unb.ca/~rolf

[Package spatstat version 1.10-3 Index]