contour.im {spatstat}	R Documentation

Contour plot of pixel image

Description

Generates a contour plot of a pixel image.

Usage


   ## S3 method for class 'im':
   contour(x, ..., main, axes=TRUE, add=FALSE)

Arguments

x 	 Pixel image to be plotted. An object of class "im".

main 	 Character string to be displayed as the main title.

axes 	 Logical. If TRUE, coordinate axes are plotted (with tick marks) around a region slightly larger than the image window. If FALSE, no axes are plotted, and a box is drawn tightly around the image window. Ignored if add=TRUE.

add 	 Logical. If FALSE, a new plot is created. If TRUE, the contours are drawn over the existing plot.

... 	 Other arguments passed to contour.default controlling the contour plot; see Details.

Details

This is a method for the generic contour function, for objects of the class "im".

An object of class "im" represents a pixel image; see im.object.

This function displays the values of the pixel image x as a contour plot on the current plot device, using equal scales on the x and y axes.

The appearance of the plot can be modified using any of the arguments listed in the help for contour.default. Useful ones include:

nlevels

Number of contour levels to plot.

drawlabels

Whether to label the contour lines with text.

col,lty,lwd

Colour, type, and width of contour lines.

See contour.default for a full list of these arguments.

The defaults for any of the abovementioned arguments can be reset using spatstat.options.

Value

none.

Author(s)

Adrian Baddeley adrian@maths.uwa.edu.au http://www.maths.uwa.edu.au/~adrian/ and Rolf Turner rolf@math.unb.ca http://www.math.unb.ca/~rolf

See Also

im.object, plot.im, persp.im

Examples


   # an image
   Z <- setcov(owin())
   contour(Z)
   contour(Z, axes=FALSE)

[Package spatstat version 1.10-3 Index]