crossdist {spatstat}	R Documentation

Pairwise distances

Description

Computes the distances between pairs of `things' taken from two different datasets.

Usage


  crossdist(X, Y, ...)

Arguments

X,Y 	 Two objects of the same class.

... 	 Additional arguments depending on the method.

Details

Given two datasets X and Y (representing either two point patterns or two line segment patterns) crossdist computes the Euclidean distance from each thing in the first dataset to each thing in the second dataset, and returns a matrix containing these distances.

The function crossdist is generic, with methods for point patterns (objects of class "ppp"), line segment patterns (objects of class "psp"), and a default method. See the documentation for crossdist.ppp, crossdist.psp or crossdist.default for further details.

Value

A matrix whose [i,j] entry is the distance from the i-th thing in the first dataset to the j-th thing in the second dataset.

Author(s)

Adrian Baddeley adrian@maths.uwa.edu.au http://www.maths.uwa.edu.au/~adrian/

See Also

crossdist.ppp, crossdist.psp, crossdist.default, pairdist, nndist

[Package spatstat version 1.10-3 Index]