crossing.psp {spatstat}	R Documentation

Crossing Points of Two Line Segment Patterns

Description

Finds any crossing points between two line segment patterns.

Usage


  crossing.psp(A,B)

Arguments

A,B 	 Line segment patterns (objects of class "psp").

Details

This function finds any crossing points between the line segment patterns A and B.

A crossing point occurs whenever one of the line segments in A intersects one of the line segments in B, at a nonzero angle of intersection.

Value

Point pattern (object of class "ppp").

Author(s)

Adrian Baddeley adrian@maths.uwa.edu.au http://www.maths.uwa.edu.au/~adrian/ and Rolf Turner rolf@math.unb.ca http://www.math.unb.ca/~rolf

See Also

selfcrossing.psp, psp.object, ppp.object.

Examples


  a <- psp(runif(10), runif(10), runif(10), runif(10), window=owin())
  b <- psp(runif(10), runif(10), runif(10), runif(10), window=owin())
  plot(a, col="green", main="crossing.psp")
  plot(b, add=TRUE, col="blue")
  P <- crossing.psp(a,b)
  plot(P, add=TRUE, col="red")

[Package spatstat version 1.10-3 Index]