cut.ppp {spatstat}	R Documentation

Convert Point Pattern Marks from Numeric to Factor

Description

Transform the marks of a marked point pattern from numeric values into a factor.

Usage


  ## S3 method for class 'ppp':
  cut(x, ...)

Arguments

x 	 A two-dimensional point pattern. An object of class "ppp". This should be a marked point pattern, and the marks would normally be numerical values.

... 	 Arguments passed to cut.default. They determine the breakpoints for the mapping from numerical values to factor values. See cut.default.

Details

This simple function applies the generic cut operation to the marks of the point pattern x. The range of values of the numerical marks is divided into several intervals, and each interval is associated with a level of a factor. The result is a marked point pattern, with the same window and point locations as x, but with the numeric mark of each point discretised by replacing it by the factor level.

This function is a convenient way to transform a marked point pattern which has numeric marks into a multitype point pattern, for example to plot it or analyse it. See the examples.

To select some points from a point pattern, use the subset operator [.ppp instead.

Value

A multitype point pattern, that is, a point pattern object (of class "ppp") with a marks vector that is a factor.

Author(s)

Adrian Baddeley adrian@maths.uwa.edu.au http://www.maths.uwa.edu.au/~adrian/ and Rolf Turner rolf@math.unb.ca http://www.math.unb.ca/~rolf

See Also

cut, ppp.object

Examples


 data(longleaf)
 # Longleaf Pines data
 # the marks are positive real numbers indicating tree diameters.

 
 ## Not run: 
 plot(longleaf)
 
## End(Not run)

 # cut the range of tree diameters into three intervals
 long3 <- cut(longleaf, 3)
 ## Not run: 
 plot(long3)
 
## End(Not run)

 # adult trees defined to have diameter at least 30 cm
 long2 <- cut(longleaf, breaks=c(0,30,100), labels=c("Sapling", "Adult"))
 plot(long2)
 plot(long2, cols=c("green","blue"))

[Package spatstat version 1.10-3 Index]