data.ppm {spatstat}	R Documentation

Extract Original Data from a Fitted Point Process Model

Description

Given a fitted point process model, this function extracts the original point pattern dataset to which the model was fitted.

Usage


  data.ppm(object)

Arguments

object 	 fitted point process model (an object of class "ppm").

Details

An object of class "ppm" represents a point process model that has been fitted to data. It is typically produced by the model-fitting algorithm ppm. The object contains complete information about the original data point pattern to which the model was fitted. This function extracts the original data pattern.

See ppm.object for a list of all operations that can be performed on objects of class "ppm".

Value

A point pattern (object of class "ppp").

Author(s)

Adrian Baddeley adrian@maths.uwa.edu.au http://www.maths.uwa.edu.au/~adrian/ and Rolf Turner rolf@math.unb.ca http://www.math.unb.ca/~rolf

See Also

ppm.object, ppp.object

Examples


 data(cells)
 fit <- ppm(cells, ~1, Strauss(r=0.1), rbord=0.1)
 X <- data.ppm(fit)
 # 'X' is identical to 'cells'

[Package spatstat version 1.10-3 Index]