default.dummy {spatstat}	R Documentation

Generate a Default Pattern of Dummy Points

Description

Generates a default pattern of dummy points for use in a quadrature scheme.

Usage


 default.dummy(X, nd, random=FALSE, ntile=NULL, npix=NULL, ..., verbose=FALSE)

Arguments

X 	 The observed data point pattern. An object of class "ppp" or in a format recognised by as.ppp()

nd 	 Optional. Integer, or integer vector of length 2, specifying an nd * nd or nd[1] * nd[2] rectangular array of dummy points.

random 	 Logical value. If TRUE, the dummy points are randomised.

ntile 	 Optional. Integer or pair of integers specifying the number of rows and columns of tiles used in the counting rule.

npix 	 Optional. Integer or pair of integers specifying the number of rows and columns of pixels used in computing approximate areas.

... 	 Ignored.

verbose 	 If TRUE, information about the construction of the quadrature scheme is printed.

Details

This function provides a sensible default for the dummy points in a quadrature scheme.

A quadrature scheme consists of the original data point pattern, an additional pattern of dummy points, and a vector of quadrature weights for all these points. See quad.object for further information about quadrature schemes.

If random is false (the default), then the function creates dummy points in an nd[1] by nd[1] rectangular grid. If random is true, then the frame of the window is divided into an nd[1] by nd[1] array of tiles, and one dummy point is generated at random inside each tile. In either case, the four corner points of the frame of the window are added. Then if the window is not rectangular, any dummy points lying outside it are deleted.

If nd is missing, a default value (depending on the data pattern X) is computed by default.ngrid.

Alternative functions for creating dummy patterns include corners, gridcentres, stratrand and spokes.

Value

A point pattern (an object of class "ppp", see ppp.object) containing the dummy points.

Author(s)

Adrian Baddeley adrian@maths.uwa.edu.au http://www.maths.uwa.edu.au/~adrian/ and Rolf Turner rolf@math.unb.ca http://www.math.unb.ca/~rolf

See Also

quad.object, quadscheme, corners, gridcentres, stratrand, spokes

Examples


  data(simdat)
  P <- simdat
  D <- default.dummy(P, 100)
  ## Not run: plot(D)
  Q <- quadscheme(P, D, "grid")
  ## Not run: plot(union.quad(Q))

[Package spatstat version 1.10-3 Index]