density.ppp {spatstat}	R Documentation

Kernel Smoothed Intensity of Point Pattern

Description

Compute a kernel smoothed intensity function from a point pattern.

Usage


  ## S3 method for class 'ppp':
  density(x, sigma, ..., weights, edge=TRUE, varcov=NULL)

Arguments

x 	 Point pattern (object of class "ppp") to be smoothed.

sigma 	 Standard deviation of isotropic Gaussian smoothing kernel.

weights 	 Optional vector of weights to be attached to the points. May include negative values.

... 	 Arguments passed to as.mask to determine the pixel resolution.

edge 	 Logical flag: if TRUE, apply edge correction.

varcov 	 Variance-covariance matrix of anisotropic Gaussian kernel. Incompatible with sigma.

Details

This is a method for the generic function density.

A kernel estimate of the intensity function of the point pattern is computed. The result is the convolution of the isotropic Gaussian kernel of standard deviation sigma with point masses at each of the data points. The default is to assign a unit weight to each point. If weights is present, the point masses have these weights (which may be signed real numbers).

If edge=TRUE, the intensity estimate is corrected for edge effect bias by dividing it by the convolution of the Gaussian kernel with the window of observation.

Instead of the isotropic Gaussian kernel with standard deviation sigma, the smoothing kernel may be chosen to be any Gaussian kernel, by giving the variance-covariance matrix varcov. The arguments sigma and varcov are incompatible. Also sigma may be a vector of length 2 giving the standard deviations of two independent Gaussian coordinates, thus equivalent to varcov = diag(sigma^2).

Computation is performed using the Fast Fourier Transform. Accuracy depends on the pixel resolution, controlled by the arguments ... passed to as.mask.

Value

A pixel image (object of class "im").

Author(s)

Adrian Baddeley adrian@maths.uwa.edu.au http://www.maths.uwa.edu.au/~adrian/ and Rolf Turner rolf@math.unb.ca http://www.math.unb.ca/~rolf

See Also

ppp.object, im.object

Examples


  data(cells)
  Z <- density.ppp(cells, 0.05)
  plot(Z)

[Package spatstat version 1.10-3 Index]