density.splitppp {spatstat}	R Documentation

Kernel Smoothed Intensity of Split Point Pattern

Description

Compute a kernel smoothed intensity function for each of the components of a split point pattern.

Usage


  ## S3 method for class 'splitppp':
  density(x, ...)

Arguments

x 	 Split point pattern (object of class "splitppp" created by split.ppp) to be smoothed.

... 	 Arguments passed to density.ppp to control the smoothing, pixel resolution, edge correction etc.

Details

This is a method for the generic function density.

The argument x should be an object of class "splitppp", effectively a list of point patterns.

Typically x is obtained by applying the function split.ppp to a point pattern y by calling split(y). This splits the points of y into several sub-patterns.

A kernel estimate of the intensity function of each of the point patterns is computed using density.ppp.

The return value is a list, each of whose entries is a pixel image (object of class "im"). The return value also belongs to the class "listof" and can be plotted or printed.

Value

A list of pixel images (objects of class "im"). Can be plotted or printed.

Author(s)

Adrian Baddeley adrian@maths.uwa.edu.au http://www.maths.uwa.edu.au/~adrian/ and Rolf Turner rolf@math.unb.ca http://www.math.unb.ca/~rolf

See Also

ppp.object, im.object

Examples


  data(amacrine)
  Z <- density(split(amacrine), 0.05)
  plot(Z)

[Package spatstat version 1.10-3 Index]