diameter {spatstat}	R Documentation

Diameter of a Window Frame

Description

Computes the diameter of a window

Usage


 diameter(w)

Arguments

w 	 A window, whose diameter will be computed.

Details

It is intended that this function compute the diameter of a window of arbitrary shape, i.e. the maximum distance between any two points in the window. The current implementation just computes the length of the diagonal of the rectangular frame surrounding the window.

The argument w should be a window (an object of class "owin", see owin.object for details) or can be given in any format acceptable to as.owin().

Value

The numerical value of the diameter of the window.

Author(s)

Adrian Baddeley adrian@maths.uwa.edu.au http://www.maths.uwa.edu.au/~adrian/ and Rolf Turner rolf@math.unb.ca http://www.math.unb.ca/~rolf

See Also

owin, as.owin

Examples


  w <- owin(c(0,1),c(0,1))
  diameter(w) 
  # returns sqrt(2)

[Package spatstat version 1.10-3 Index]