distmap.psp {spatstat}	R Documentation

Distance Map of Line Segment Pattern

Description

Computes the distance from each pixel to the nearest line segment in the given line segment pattern.

Usage


  ## S3 method for class 'psp':
  distmap(X, ...)

Arguments

X 	 A line segment pattern (object of class "psp").

... 	 Arguments passed to as.mask to control pixel resolution.

Details

The ``distance map'' of a line segment pattern X is the function f whose value f(u) is defined for any two-dimensional location u as the shortest distance from u to X.

This function computes the distance map of the line segment pattern X and returns the distance map as a pixel image. The greyscale value at a pixel u equals the distance from u to the nearest line segment of the pattern X.

Additionally, the return value has two attributes, "index" and "bdry", which are also pixel images. The grey values in "bdry" give the distance from each pixel to the bounding rectangle of the image. The grey values in "index" are integers identifying which line segment of X is closest.

This is a method for the generic function distmap.

Value

A pixel image (object of class "im") whose greyscale values are the values of the distance map. The return value has attributes "index" and "bdry" which are also pixel images.

Author(s)

Adrian Baddeley adrian@maths.uwa.edu.au http://www.maths.uwa.edu.au/~adrian/ and Rolf Turner rolf@math.unb.ca http://www.math.unb.ca/~rolf

See Also

distmap, distmap.owin, distmap.ppp

Examples


    a <- psp(runif(20),runif(20),runif(20),runif(20), window=owin())
    Z <- distmap(a)
    plot(Z)
    plot(a, add=TRUE)

[Package spatstat version 1.10-3 Index]