dummy.ppm {spatstat}	R Documentation

Extract Dummy Points Used to Fit a Point Process Model

Description

Given a fitted point process model, this function extracts the `dummy points' of the quadrature scheme used to fit the model.

Usage


  dummy.ppm(object)

Arguments

object 	 fitted point process model (an object of class "ppm").

Details

An object of class "ppm" represents a point process model that has been fitted to data. It is typically produced by the model-fitting algorithm ppm.

The ppm algorithm approximates the pseudolikelihood integral by a sum over a finite set of quadrature points, which is constructed by augmenting the original data point pattern by a set of ``dummy'' points. The fitted model object returned by ppm contains complete information about this quadrature scheme. See ppm or ppm.object for further information.

This function extracts the dummy points of the quadrature scheme. A typical use of this function would be to count the number of dummy points, to gauge the accuracy of the approximation to the exact pseudolikelihood.

See ppm.object for a list of all operations that can be performed on objects of class "ppm".

Value

A point pattern (object of class "ppp").

Author(s)

Adrian Baddeley adrian@maths.uwa.edu.au http://www.maths.uwa.edu.au/~adrian/ and Rolf Turner rolf@math.unb.ca http://www.math.unb.ca/~rolf

See Also

ppm.object, ppp.object, ppm

Examples


 data(cells)
 fit <- ppm(cells, ~1, Strauss(r=0.1), rbord=0.1)
 X <- dummy.ppm(fit)
 X$n
 # this is the number of dummy points in the quadrature scheme

[Package spatstat version 1.10-3 Index]