duplicated.ppp {spatstat}	R Documentation

Determine Duplicated Points in a Spatial Point Pattern

Description

Determines which points in a spatial point pattern are duplicates of previous points, and returns a logical vector.

Usage


 ## S3 method for class 'ppp':
 duplicated(x, ...)

Arguments

x 	 A spatial point pattern (object of class "ppp").

... 	 Ignored.

Details

This is a method for the generic function duplicated for point pattern datasets (of class "ppp", see ppp.object).

Two points in a point pattern are deemed to be identical if their x,y coordinates are the same, and their marks are also the same (if they carry marks). The Examples section illustrates how it is possible for a point pattern to contain a pair of identical points.

This function determines which points in x duplicate other points that appeared earlier in the sequence. It returns a logical vector with entries that are TRUE for duplicated points and FALSE for unique (non-duplicated) points.

Value

A logical vector of length equal to the number of points in x.

Author(s)

Adrian Baddeley adrian@maths.uwa.edu.au http://www.maths.uwa.edu.au/~adrian/ and Rolf Turner rolf@math.unb.ca http://www.math.unb.ca/~rolf

See Also

ppp.object, unique.ppp, multiplicity.ppp

Examples


   X <- ppp(c(1,1,0.5), c(2,2,1), window=square(3))
   duplicated(X)

[Package spatstat version 1.10-3 Index]