erode.owin {spatstat}	R Documentation

Erode a Window

Description

Perform morphological erosion of a window

Usage


 erode.owin(w, r, shrink.frame=TRUE, ...)

Arguments

w 	 A window (object of class "owin".

r 	 positive number: the radius of erosion.

shrink.frame 	 logical: if TRUE, erode the bounding rectangle as well.

... 	 extra arguments to as.mask controlling the pixel resolution

Details

The morphological erosion of a set W by a distance r > 0 is the subset consisting of all points x in W such that the distance from x to the boundary of W is greater than or equal to r. In other words it is the result of trimming a margin of width r off the set W.

This function computes the erosion of the window w. If w is not a rectangle, it must be approximated by a binary pixel image, and the arguments "..." are passed to as.mask to determine the pixel resolution. There is a sensible default.

If shrink.frame is false, the resulting window is given the same outer, bounding rectangle as the original window w. If shrink.frame is true, the original bounding rectangle is also eroded by the same distance r.

To simply compute the area of the eroded window, use eroded.areas.

Value

Another object of class "owin" representing the eroded window.

Author(s)

Adrian Baddeley adrian@maths.uwa.edu.au http://www.maths.uwa.edu.au/~adrian/ and Rolf Turner rolf@math.unb.ca http://www.math.unb.ca/~rolf

See Also

owin, as.owin, eroded.areas

Examples


  w <- owin(c(0,1),c(0,1))
  v <- erode.owin(w, 0.1) 
  # returns rectangle [0.1, 0.9] x [0.1,0.9]
  ## Not run: 
  v <- erode.owin(w, 0.6)
  #fails - erosion is empty
  
## End(Not run)

[Package spatstat version 1.10-3 Index]