expand.owin {spatstat}	R Documentation

Expand Window By Factor

Description

Expands the bounding box of a window by a given factor.

Usage


 expand.owin(W, f=1)

Arguments

W 	 A window.

f 	 Scalar factor for expansion of the area of the window.

Details

The argument w should be a window (an object of class "owin", see owin.object for details) or can be given in any format acceptable to as.owin().

The bounding box of the window W is expanded by a scale factor equal to the square root of f in both the x and y directions, so that the area increases by a factor f. The centre of this expanded rectangle is the same as the centre of the original bounding box. The expanded rectangle is returned.

This quirky little function exists mainly for reference, as it is the rule for calculating window expansion in our implementation of the Metropolis-Hastings algorithm rmh.default.

For general transformations of the scale, location and orientation of a window, see shift, affine and rotate.

Value

A rectangular window.

Author(s)

Adrian Baddeley adrian@maths.uwa.edu.au http://www.maths.uwa.edu.au/~adrian/ and Rolf Turner rolf@math.unb.ca http://www.math.unb.ca/~rolf

See Also

owin, as.owin, shift, rotate, affine

Examples


  w <- square(1)
  expand.owin(w, 9)
  # returns the square [-1,2] x [-1,2]

[Package spatstat version 1.10-3 Index]