fitted.ppm {spatstat}	R Documentation

Fitted Conditional Intensity for Point Process Model

Description

Given a point process model fitted to a point pattern, compute the fitted conditional intensity of the model at the points of the pattern, or at the points of the quadrature scheme used to fit the model.

Usage


  ## S3 method for class 'ppm':
  fitted(object, ..., type="lambda", dataonly=FALSE,
check=TRUE, repair=TRUE)

Arguments

object 	 The fitted point process model (an object of class "ppm")

... 	 Ignored.

type 	 String (partially matched) indicating whether the fitted value is the conditional intensity ("lambda") or the trend ("trend").

dataonly 	 Logical. If TRUE, then values will only be computed at the points of the data point pattern. If FALSE, then values will be computed at all the points of the quadrature scheme used to fit the model, including the points of the data point pattern.

check 	 Logical value indicating whether to check the internal format of object. If there is any possibility that this object has been restored from a dump file, or has otherwise lost track of the environment where it was originally computed, set check=TRUE.

repair 	 Logical value indicating whether to repair the internal format of object, if it is found to be damaged.

Details

The argument object must be a fitted point process model (object of class "ppm"). Such objects are produced by the model-fitting algorithm ppm).

This function evaluates the conditional intensity lambdahat(u,x) or spatial trend bhat(u) of the fitted point process model for certain locations u, where x is the original point pattern dataset to which the model was fitted.

The locations u at which the fitted conditional intensity/trend is evaluated, are the points of the quadrature scheme used to fit the model in ppm. They include the data points (the points of the original point pattern dataset x) and other ``dummy'' points in the window of observation.

Use predict.ppm to compute the fitted conditional intensity at other locations or with other values of the explanatory variables.

Value

A vector containing the values of the fitted conditional intensity or (if type="trend") the fitted spatial trend.
Entries in this vector correspond to the quadrature points (data or dummy points) used to fit the model. The quadrature points can be extracted from object by union.quad(quad.ppm(object)).

Author(s)

Adrian Baddeley adrian@maths.uwa.edu.au http://www.maths.uwa.edu.au/~adrian/ and Rolf Turner rolf@math.unb.ca http://www.math.unb.ca/~rolf

References

Baddeley, A., Turner, R., Moller, J. and Hazelton, M. (2005). Residual analysis for spatial point processes (with discussion). Journal of the Royal Statistical Society, Series B 67, 617–666.

See Also

ppm.object, ppm, predict.ppm

Examples


    data(cells)
    str <- ppm(cells, ~x, Strauss(r=0.15), rbord=0.15)
    lambda <- fitted(str)

    # extract quadrature points in corresponding order
    quadpoints <- union.quad(quad.ppm(str))

    # plot conditional intensity values
    # as circles centred on the quadrature points 
    quadmarked <- setmarks(quadpoints, lambda)
    plot(quadmarked)

[Package spatstat version 1.10-3 Index]