foo {spatstat}	R Documentation

Foo is Not a Real Name

Description

The name foo is not a real name: it is a place holder, used to represent the name of any desired thing.

The functions defined here simply print an explanation of the placeholder name foo.

Usage

foo()

## S3 method for class 'foo'
plot(x, ...)

Arguments

x
Ignored.

...
Ignored.

Details

The name foo is used by computer scientists as a place holder, to represent the name of any desired object or function. It is not the name of an actual object or function; it serves only as an example, to explain a concept.

However, many users misinterpret this convention, and actually type the command foo or foo(). Then they email the package author to inform them that foo is not defined.

To avoid this correspondence, we have now defined an object called foo.

The function foo() prints a message explaining that foo is not really the name of a variable.

The function can be executed simply by typing foo without parentheses.

Value

Null.

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au, Rolf Turner rolfturner@posteo.net and Ege Rubak rubak@math.aau.dk.

See Also

beginner

Examples

  foo

[Package spatstat version 3.3-1 Index]