fv.object {spatstat}	R Documentation

Data Frames of Function Values

Description

A class "fv" to support the convenient plotting of several estimates of the same function.

Details

An object of this class is a convenient way of storing and plotting several different estimates of the same function.

It is a data frame with extra attributes indicating the recommended way of plotting the function, and other information.

There are methods for print and plot for this class.

Objects of class "fv" are returned by Fest, Gest,Jest, and Kest.

Author(s)

Adrian Baddeley adrian@maths.uwa.edu.au http://www.maths.uwa.edu.au/~adrian/ and Rolf Turner rolf@math.unb.ca http://www.math.unb.ca/~rolf

See Also

Fest, Gest, Jest, Kest

Examples


    data(cells)
    K <- Kest(cells)

    class(K)

    K  # prints a sensible summary

    plot(K)

[Package spatstat version 1.10-3 Index]