identify.ppp {spatstat}	R Documentation

Identify Points in a Point Pattern

Description

If a point pattern is plotted in the graphics window, this function will find the point of the pattern which is nearest to the mouse position, and print its mark value (or its serial number if there is no mark).

Usage


  ## S3 method for class 'ppp':
  identify(x, ...)

Arguments

x 	 A point pattern (object of class "ppp").

... 	 Arguments passed to identify.default.

Details

This is a method for the generic function identify for point pattern objects.

The point pattern x should first be plotted using plot.ppp. Then identify(x) reads the position of the graphics pointer each time the left mouse button is pressed. It then finds the point of the pattern x closest to the mouse position. If this closest point is sufficiently close to the mouse pointer, its index (and its mark if any) will be returned as part of the value of the call.

Each time a point of the pattern is identified, text will be displayed next to the point, showing its serial number (if x is unmarked) or its mark value (if x is marked).

Value

If x is unmarked, the result is a vector containing the serial numbers of the points in the pattern x that were identified. If x is marked, the result is a 2-column matrix, the first column containing the serial numbers and the second containing the marks for these points.

Author(s)

Adrian Baddeley adrian@maths.uwa.edu.au http://www.maths.uwa.edu.au/~adrian/ and Rolf Turner rolf@math.unb.ca http://www.math.unb.ca/~rolf

See Also

identify, clickppp

[Package spatstat version 1.10-3 Index]