spatstat-internal {spatstat}	R Documentation

Internal spatstat functions

Description

Internal spatstat functions.

Usage


affinexy(X, mat = diag(c(1, 1)), vec = c(0, 0))
apply23sum(x)
area.xypolygon(polly)
AsymmHausdorff.psp(X, Y, method="Fortran")
as.breakpts(...)
as.fv(x)
as.polygonal(W)
as.units(s)
beachcolours(heightrange, sealevel, monochrome, ncolours)
bind.fv(x, y, labl, desc, preferred)
bdrylength.xypolygon(polly)
breakpts(val, maxi, even = FALSE, npos = NULL, step = NULL)
breakpts.from.r(r)
cartesian(pp, markset, fac = TRUE)
cellmiddles(W, nx, ny, npix, gi)
checkfields(X,L)          
check.hist.lengths(hist,breaks)
check.named.vector(x, nam)
clip.psp(x, window, check=TRUE)
cliprect.psp(x, window)
clippoly.psp(s, window)
closepairs(X,rmax)
crosspairs(X,Y,rmax)
cobble.xy(x, y, f, fatal)
commasep(x)
countingweights(id, areas, check = TRUE)
damaged.ppm(object)
default.expand(object, m, epsilon)
default.clipwindow(object, epsilon)
default.n.tiling(X, nd, ntile, npix, verbose)
default.ntile(X)
densityhack(x, bw, adjust, kernel, window, width,
             give.Rkern, n, from, to, cut, na.rm, weights)
diagnose.ppm.engine(object, ..., type="eem", typename, opt,
                         sigma=NULL, rbord = reach(object), compute.sd=TRUE,
                         compute.cts=TRUE, rv=NULL, oldstyle=FALSE)
discretise(x, xrange, nx)
distpl(p, l)               
distppl(p, l)
distppll(p, l, mintype=0, method="Fortran", listit=FALSE)
divisors(n)
do.call.matched(fname, arglist, funargs, extrargs)
edge.Ripley(X, r, W, method)
edge.Trans(X, Y, W, exact, paired, trim)
ensure2vector(x)
equals.quad(Q)          
eratosthenes(nmax)
erode.mask(w,r)
even.breaks.owin(w)
exactdt(X, ...)              
exactPdt(w)
fasp(fns, titles, formulae, which, dataname, title)
findmarktype(x)
FormatFaspFormulae(f, argname)
fv(x, argu, ylab, valu, fmla, alim, labl, desc, units)
greatest.common.divisor(n,m)
getfields(X, L, fatal = TRUE)
getglmfit(object)
gridindex(x, y, xrange, yrange, nx, ny)            
handle.r.b.args(r = NULL, breaks = NULL, window, eps = NULL, rmaxdefault)
handle.rshift.args(W, ..., radius, width, height, edge, clip, edgedefault)
ho.engine(model, ..., nsim, nrmh, start, control, verb)
identical.formulae(x,y)
inside.xypolygon(pts, polly, test01 = TRUE, method="Fortran")     
intX.owin(w)
intX.xypolygon(polly)
intY.owin(w)
intY.xypolygon(polly)
is.cadlag(s)
is.data(Q)
is.fv(x)
is.marked.default(...)  
is.marked.quad(X, na.action="warn", ...)
is.multitype.default(...)  
is.multitype.quad(X, na.action="warn", ...)
is.poisson.ppm(x)
is.prime(n)
is.stationary.ppm(x)    
km.rs(o, cc, d, breaks)
Kborder.engine(X, rmax, nr, correction, weights)
Kount(dIJ, bI, b, breaks)
Kwtsum(dIJ, bI, wIJ, b, w, breaks)
Kmulti.inhom(X, I, J, lambdaI, lambdaJ, ...,
         r=NULL, breaks=NULL,
         correction = c("border", "isotropic", "Ripley", "translate") ,
         lambdaIJ=NULL,
         Iname = "points satisfying condition I",
         Jname = "points satisfying condition J")
killinteraction(model)
least.common.multiple(n,m)
## S3 method for class 'im':
levels(x) <- value
lookup.im(Z, x, y, naok)
make.even.breaks(bmax, npos, bstep)
makeunits(sing, plur, mul)
markformat(x)
## S3 method for class 'ppp':
markformat(x)
## S3 method for class 'psp':
markformat(x)
markspace.integral(X)
marks(x, ...)
marks.ppp(x, ..., dfok)           
marks.psp(x, ..., dfok)
marks.quad(x, dfok=FALSE, ...)
marks(x, ...) <- value
## S3 method for class 'ppp':
marks(x, dfok, ...) <- value
markappendop(x, y)
marksubset(x, index, format)
marksubsetop(x, i)
markreplicateop(x, n)
matcolall(x)
matcolany(x)
matcolsum(x)            
matrixsample(mat, newdim, phase)
matrowall(x)
matrowany(x)
matrowsum(x)
maxflow(costm)
meanX.owin(w)            
meanY.owin(w)
mkcor(d, ff, wt, Ef, rvals, method="smrep", ..., nwtsteps=500)
mpl.engine(Q, trend, interaction, ..., covariates, correction,
         rbord, use.gam, famille,
         forcefit, allcovar, callstring, precomputed, savecomputed, preponly)
mpl.get.covariates(covariates, locations, type)
mpl.prepare(Q, X, P, trend, interaction, covariates, 
            want.trend, want.inter, correction, rbord, Pname,
            callstring, ..., allcovar, precomputed, savecomputed)
MultiPair.checkmatrix(mat, n, name)
nearest.pixel(x, y, im)
no.trend.ppm(x)
n.quad(Q)               
overlap.owin(A,B)
overlap.trapezium(xa, ya, xb, yb, verb = FALSE)
overlap.xypolygon(P, Q)
owinpolycheck(W, verbose=TRUE)
param.quad(Q)
pixellate(x, ..., weights)
ploterodewin(W1, W2, col.edge, col.inside, ...)
ploterodeimage(W, Z, ..., Wcol, rangeZ, colsZ)
plot.minconfit(x, ...)
plot.pppmatching(x, addmatch = NULL, main = NULL, ...)
plot.profilepl(x, ..., add=FALSE, main=NULL, tag=TRUE, coeff=NULL, xvariable=NULL)
plot.qqppm(x, ..., limits=TRUE, monochrome=FALSE,
           limcol=if(monochrome) "black" else "red")
plot.quadratcount(x, ..., add, entries, dx, dy)
plot.quadrattest(x, ...)
polynom(x, ...)
pppdist.prohorov(X, Y, precision = 7)
pppmatching(X, Y, am, ty = "generic")
primefactors(n, prmax)
## S3 method for class 'diagppm':
print(x, ...)
## S3 method for class 'envelope':
print(x, ...)
## S3 method for class 'fasp':
print(x, ...)
## S3 method for class 'fv':
print(x, ...)
## S3 method for class 'interact':
print(x, ...)       
## S3 method for class 'isf':
print(x, ...)
## S3 method for class 'minconfit':
print(x, ...)
## S3 method for class 'plotppm':
print(x, ...)
## S3 method for class 'pppmatching':
print(x, ...)
## S3 method for class 'profilepl':
print(x, ...)
## S3 method for class 'qqppm':
print(x, ...)
## S3 method for class 'rmhcontrol':
print(x, ...)
## S3 method for class 'rmhmodel':
print(x, ...)
## S3 method for class 'rmhstart':
print(x, ...)
## S3 method for class 'rmhseed':
print(x, ...)
## S3 method for class 'summary.owin':
print(x, ...)
## S3 method for class 'summary.ppp':
print(x, ..., dp=3)
## S3 method for class 'summary.psp':
print(x, ...)
## S3 method for class 'summary.units':
print(x, ...)
## S3 method for class 'units':
print(x, ...)
quad(data, dummy, w, param)
quadrat.breaks(xr, yr, nx = 5, ny = nx, xbreaks = NULL, ybreaks = NULL)
quadrat.count.engine(x, y, xbreaks, ybreaks, weights)
quadscheme.replicated(data, dummy, method = "grid", ...)
quadscheme.spatial(data, dummy, method = "grid", ...)
rasterx.im(x)
rastery.im(x)
rebound.ppp(x, ...)
rebound.psp(x, ...)
rebound.owin(w, rect)
## S3 method for class 'units':
rescale(X,s)
resid4plot(RES, plot.neg="image", plot.smooth="imagecontour",
           spacing=0.1, srange=NULL,monochrome=FALSE, main=NULL, ...)
resid1plot(RES, opt, plot.neg="image", plot.smooth="imagecontour",
              srange=NULL, monochrome=FALSE, main=NULL, ...)
resid1panel(observedX, observedV,
            theoreticalX, theoreticalV, theoreticalSD, xlab,ylab, ...)
resolve.defaults(...)
reverse.xypolygon(p)
revcumsum(x)
rmax.rule(fun, W, lambda)
rotxy(X, angle = pi/2)
rmhResolveTypes(model, start, control)
rmhcontrol.rmhcontrol(control, ...)
rmhcontrol.list(control, ...)
rmhEngine(InfoList, ..., verbose, reseed, kitchensink, preponly) 
rmhmodel.rmhmodel(model, ...)
rmhmodel.list(model, ...)
rmhseed(iseed)
rmhstart.rmhstart(start, ...)
rmhstart.list(start, ...)
rmhmodel.ppm(model, win, ..., verbose, project, control)
rmpoint.I.allim(n, f, types)
rpoint.multi(n, f, fmax=NULL, marks = NULL, win = unit.square(), giveup = 1000, verbose = FALSE)
runifpoispp(lambda, win = owin(c(0, 1), c(0, 1)))
runifrect(n, win = owin(c(0, 1), c(0, 1)))
second.moment.calc(x, sigma, edge=TRUE, what="Kmeasure", debug=FALSE,
..., varcov=NULL)
shiftxy(X, vec = c(0, 0))
spatstat.rawdata.location(...)             
sp.foundclass(cname, inlist, formalname, argsgiven)             
sp.foundclasses(cnames, inlist, formalname, argsgiven)             
stratrand(window, nx, ny, k = 1)
suffloc(object)
suffstat.generic(model, X, callstring)
suffstat.poisson(model, X, callstring)
## S3 method for class 'envelope':
summary(object,...)
## S3 method for class 'profilepl':
summary(object, ...)
## S3 method for class 'pppmatching':
summary(object, ...)
## S3 method for class 'units':
summary(object, ...)
sympoly(x, y, n)
termsinformula(x)
tilecentroids(W, nx, ny)        
## Default S3 method:
units(x)
## Default S3 method:
units(x) <- value
update.interact(object, ...)
validate.mask(w, fatal=TRUE)        
validate.quad(Q, fatal, repair, announce)
variablesinformula(x)
verifyclass(X, C, N = deparse(substitute(X)), fatal = TRUE)
verify.xypolygon(p, fatal=TRUE)
whist(x,breaks,weights, trim, right)
w.quad(Q)               
x.quad(Q)
y.quad(Q)
xypolyselfint(p, eps, proper)
xypolygon2psp(p, w)

Details

These are usually not to be called by the user.

[Package spatstat version 1.10-3 Index]