is.im {spatstat}	R Documentation

Test Whether An Object Is A Pixel Image

Description

Tests whether its argument is a pixel image (object of class "im").

Usage


is.im(x)

Arguments

x 	 Any object.

Details

This function tests whether the argument x is a pixel image object of class "im". For details of this class, see im.object.

The object is determined to be an image if it inherits from class "im".

Value

TRUE if x is a pixel image, otherwise FALSE.

Author(s)

Adrian Baddeley adrian@maths.uwa.edu.au http://www.maths.uwa.edu.au/~adrian/ and Rolf Turner rolf@math.unb.ca http://www.math.unb.ca/~rolf

[Package spatstat version 1.10-3 Index]