is.marked.ppm {spatstat}	R Documentation

Test Whether A Point Process Model is Marked

Description

Tests whether a fitted point process model involves ``marks'' attached to the points.

Usage


  ## S3 method for class 'ppm':
  is.marked(X, ...) 

Arguments

X 	 Fitted point process model (object of class "ppm") usually obtained from ppm.

... 	 Ignored.

Details

``Marks'' are observations attached to each point of a point pattern. For example the longleaf dataset contains the locations of trees, each tree being marked by its diameter; the amacrine dataset gives the locations of cells of two types (on/off) and the type of cell may be regarded as a mark attached to the location of the cell.

The argument X is a fitted point process model (an object of class "ppm") typically obtained by fitting a model to point pattern data using ppm.

This function returns TRUE if the original data (to which the model X was fitted) were a marked point pattern.

Note that this is not the same as testing whether the model involves terms that depend on the marks (i.e. whether the fitted model ignores the marks in the data). Currently we have not implemented a test for this.

If this function returns TRUE, the implications are (for example) that any simulation of this model will require simulation of random marks as well as random point locations.

Value

Logical value, equal to TRUE if X is a model that was fitted to a marked point pattern dataset.

Author(s)

Adrian Baddeley adrian@maths.uwa.edu.au http://www.maths.uwa.edu.au/~adrian/ and Rolf Turner rolf@math.unb.ca http://www.math.unb.ca/~rolf

See Also

is.marked, is.marked.ppp

Examples


   data(lansing)
   # Multitype point pattern --- trees marked by species

   

  fit1 <- ppm(lansing, ~ marks, Poisson())
  is.marked(fit1)
  # TRUE

  fit2 <- ppm(lansing, ~ 1, Poisson())
  is.marked(fit2)
  # TRUE

  data(cells)
  # Unmarked point pattern

  fit3 <- ppm(cells, ~ 1, Poisson())
  is.marked(fit3)
  # FALSE


[Package spatstat version 1.10-3 Index]