is.marked.ppp {spatstat}	R Documentation

Test Whether A Point Pattern is Marked

Description

Tests whether a point pattern has ``marks'' attached to the points.

Usage


  ## S3 method for class 'ppp':
  is.marked(X, na.action="warn", ...) 

Arguments

X 	 Point pattern (object of class "ppp")

na.action 	 String indicating what to do if NA values are encountered amongst the marks. Options are "warn", "fatal" and "ignore".

... 	 Ignored.

Details

``Marks'' are observations attached to each point of a point pattern. For example the longleaf dataset contains the locations of trees, each tree being marked by its diameter; the amacrine dataset gives the locations of cells of two types (on/off) and the type of cell may be regarded as a mark attached to the location of the cell.

This function tests whether the point pattern X contains or involves marked points. It is a method for the generic function is.marked.

The argument na.action determines what action will be taken if the point pattern has a vector of marks but some or all of the marks are NA. Options are "fatal" to cause a fatal error; "warn" to issue a warning and then return TRUE; and "ignore" to take no action except returning TRUE.

Value

Logical value, equal to TRUE if X is a marked point pattern.

Author(s)

Adrian Baddeley adrian@maths.uwa.edu.au http://www.maths.uwa.edu.au/~adrian/ and Rolf Turner rolf@math.unb.ca http://www.math.unb.ca/~rolf

See Also

is.marked, is.marked.ppm

Examples


   data(cells)
   is.marked(cells)  #FALSE
   data(longleaf)
   is.marked(longleaf) #TRUE

[Package spatstat version 1.10-3 Index]