is.multitype {spatstat}	R Documentation

Test whether Object is Multitype

Description

Generic function to test whether a given object (usually a point pattern or something related to a point pattern) has ``marks'' attached to the points which classify the points into several types.

Usage


  is.multitype(X, ...) 

Arguments

X 	 Object to be inspected

... 	 Other arguments.

Details

``Marks'' are observations attached to each point of a point pattern. For example the longleaf dataset contains the locations of trees, each tree being marked by its diameter; the amacrine dataset gives the locations of cells of two types (on/off) and the type of cell may be regarded as a mark attached to the location of the cell. Other objects related to point patterns, such as point process models, may involve marked points.

This function tests whether the object X contains or involves marked points, and that the marks are a factor.

For example, the amacrine dataset is multitype (there are two types of cells, on and off), but the longleaf dataset is not multitype (the marks are real numbers).

This function is generic; methods are provided for point patterns (objects of class "ppp") and point process models (objects of class "ppm").

Value

Logical value, equal to TRUE if X is multitype.

Author(s)

Adrian Baddeley adrian@maths.uwa.edu.au http://www.maths.uwa.edu.au/~adrian/ and Rolf Turner rolf@math.unb.ca http://www.math.unb.ca/~rolf

See Also

is.multitype.ppp, is.multitype.ppm

[Package spatstat version 1.10-3 Index]