is.ppp {spatstat}	R Documentation

Test Whether An Object Is A Point Pattern

Description

Checks whether its argument is a point pattern (object of class "ppp").

Usage


is.ppp(x)

Arguments

x 	 Any object.

Details

This function tests whether the object x is a point pattern object of class "ppp". See ppm.object for details of this class.

The result is determined to be TRUE if x inherits from "ppp", i.e. if x has "ppp" amongst its classes.

Value

TRUE if x is a point pattern, otherwise FALSE.

Author(s)

Adrian Baddeley adrian@maths.uwa.edu.au http://www.maths.uwa.edu.au/~adrian/ and Rolf Turner rolf@math.unb.ca http://www.math.unb.ca/~rolf

[Package spatstat version 1.10-3 Index]