lansing {spatstat}	R Documentation

Lansing Woods Point Pattern

Description

Locations and botanical classification of trees in Lansing Woods.

The data come from an investigation of a 924 ft x 924 ft (19.6 acre) plot in Lansing Woods, Clinton County, Michigan USA by D.J. Gerrard. The data give the locations of 2251 trees and their botanical classification (into hickories, maples, red oaks, white oaks, black oaks and miscellaneous trees). The original plot size (924 x 924 feet) has been rescaled to the unit square.

Usage

data(lansing)

Format

An object of class "ppp" representing the point pattern of tree locations. Entries include

x 	 Cartesian x-coordinate of tree

y 	 Cartesian y-coordinate of tree

marks 	 factor with levels indicating species of each tree

The levels of marks are blackoak, hickory, maple, misc, redoak and whiteoak. See ppp.object for details of the format of a point pattern object.

Notes

Source

References

Besag, J. (1978) Some methods of statistical analysis for spatial data. Bull. Internat. Statist. Inst. 44, 77–92.

Cox, T.F. (1976) The robust estimation of the density of a forest stand using a new conditioned distance method. Biometrika 63, 493–500.

Cox, T.F. (1979) A method for mapping the dense and sparse regions of a forest stand. Applied Statistics 28, 14–19.

Cox, T.F. and Lewis, T. (1976) A conditioned distance ratio method for analysing spatial patterns. Biometrika 63, 483–492.

Diggle, P.J. (1979a) The detection of random heterogeneity in plant populations. Biometrics 33, 390–394.

Diggle, P.J. (1979b) Statistical methods for spatial point patterns in ecology. Spatial and temporal analysis in ecology. R.M. Cormack and J.K. Ord (eds.) Fairland: International Co-operative Publishing House. pages 95–150.

Diggle, P.J. (1981) Some graphical methods in the analysis of spatial point patterns. In Interpreting Multivariate Data. V. Barnett (eds.) John Wiley and Sons. Pages 55–73.

Diggle, P.J. (1983) Statistical analysis of spatial point patterns. Academic Press.

Gerrard, D.J. (1969) Competition quotient: a new measure of the competition affecting individual forest trees. Research Bulletin 20, Agricultural Experiment Station, Michigan State University.

Lotwick, H.W. (1981) Spatial stochastic point processes. PhD thesis, University of Bath, UK.

Ord, J.K. (1978) How many trees in a forest? Mathematical Scientist 3, 23–33.

[Package spatstat version 1.10-3 Index]