latest.changes {spatstat}	R Documentation

List Recent Significant Changes to a Function

Description

List all the changes to a particular function in a package, starting from a certain date or version of the package. The default is to list changes in the latest version of the spatstat family of packages.

Usage

  latest.changes(x,
          sinceversion = NULL, sincedate = NULL,
          package = spatstat.family(), show = TRUE)

Arguments

x
Character string giving the name of the function of interest, or a search pattern to be matched. A character vector is permitted.

sinceversion
Earliest version of package for which changes should be listed. A character string. Default is the most recent version.

sincedate
Earliest release date of package for which changes should be listed. A character string or a date-time object. Default is the date of the most recent version.

package
The name of the package (or packages) for which changes are to be listed. A character string or character vector.

show
Logical value indicating whether to display the table of changes on the terminal.

Details

Details of changes are extracted from the NEWS file of the specified package under the heading ‘Significant User-Visible Changes’. All entries for which the first line matches x are selected. The table of changes is displayed (if show=TRUE) and returned invisibly.

The argument sinceversion should be a character string like "1.2-3". The default is the version string of the most recent version.

The argument sincedate should be a character string like "2015-05-27", or a date-time object. The default is the date of the most recent version.

If sinceversion="all" or sincedate="all" then all recorded changes will be listed.

The special options sinceversion="book" and sincedate="book" are interpreted to mean sincedate="2015-06-05", which gives all changes reported after publication of the book by Baddeley, Rubak and Turner (2015).

By default, changes in the extension packages spatstat.local, spatstat.Knet, spatstat.gui are not reported. To include these changes as well, set package=spatstat.family(TRUE,TRUE).

Value

(Invisibly) a data frame, belonging to the class "changetable", which has a print method.

Pattern matching

The first argument x is interpreted as a pattern to be matched using grepl.

For example if x = "ppp", then this will match any news items that are about ppp, but will also match any news items about plot.ppp, print.ppp, ppplist, pppmatch, [.ppp and so on.

The pattern x = "^ppp" will match only those strings which begin with "ppp", such as ppp, ppplist, pppmatch. The symbol ^ means the start of the string.

The pattern x = "^ppp$" will match only the string "ppp". The symbol $ means the end of the string.

The pattern x = "\.ppp$" will match only those strings which end with ".ppp", such as plot.ppp, print.ppp, [.ppp. The symbol . has a special meaning, but here we want the character ".", so we use \ to escape the special meaning, so that \. means the character ".", and "\.ppp$" means any string ending in ".ppp".

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au.

References

Baddeley, A., Rubak, E. and Turner, R. (2015) Spatial Point Patterns: Methodology and Applications with R. Chapman and Hall/CRC Press.

See Also

bugfixes, latest.news, news.

Examples

   latest.changes("plot.symbolmap")

[Package spatstat version 3.3-1 Index]