latest.news {spatstat}	R Documentation

Print News About Latest Version of Package

Description

Prints the news documentation for the current version of spatstat or another specified package.

Usage

latest.news(package = spatstat.family(), doBrowse=FALSE, major=TRUE)

Arguments

package
Name of package for which the latest news should be printed. A character string, or vector of character strings.

doBrowse
Logical value indicating whether to display the results in a browser window instead of printing them.

major
Logical value. If TRUE (the default), print all information for the current major version "x.y". If FALSE, print only the information for the current minor version "x.y-z".

Details

This function prints the news documentation about changes in the current installed version of a package.

By default, it prints the latest news about all the sub-packages in the spatstat family.

The function can be called simply by typing its name without parentheses (see the Examples).

If major=FALSE, only information for the current minor version "x.y-z" will be printed. If major=TRUE (the default), all information for the current major version "x.y" will be printed, encompassing versions "x.y-0", "x.y-1", up to "x.y-z".

If package is given, then the function reads the news for the specified package from its NEWS file (if it has one) and prints only the entries that refer to the current version of the package.

To see the news for all previous versions as well as the current version, use the R utility news. See the Examples.

Value

Null.

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au, Rolf Turner rolfturner@posteo.net and Ege Rubak rubak@math.aau.dk.

See Also

spatstat.family lists the packages in the spatstat family.

bugfixes lists bug fixes.

news

Examples

  if(interactive()) {

    # current news
    latest.news

    # all news
    # news(package="spatstat")

  }

[Package spatstat version 3.3-1 Index]