lengths.psp {spatstat}	R Documentation

Lengths of Line Segments

Description

Computes the length of each line segment in a line segment pattern.

Usage


  lengths.psp(x)

Arguments

x 	 A line segment pattern (object of class "psp").

Details

The length of each line segment is computed and the lengths are returned as a numeric vector.

Value

Numeric vector.

Author(s)

Adrian Baddeley adrian@maths.uwa.edu.au http://www.maths.uwa.edu.au/~adrian/ and Rolf Turner rolf@math.unb.ca http://www.math.unb.ca/~rolf

See Also

summary.psp, midpoints.psp, angles.psp

Examples


  a <- psp(runif(10), runif(10), runif(10), runif(10), window=owin())
  b <- lengths.psp(a)   

[Package spatstat version 1.10-3 Index]