letterR {spatstat}	R Documentation

Window in Shape of Letter R

Description

A window in the shape of the capital letter R, for use in demonstrations.

Usage


 data(letterR)

Format

An object of class "owin" representing the capital letter R, in the same font as the R package logo. See owin.object for details of the format.

Source

Adrian Baddeley

[Package spatstat version 1.10-3 Index]