lgcp.estK {spatstat}	R Documentation

Fit a Log-Gaussian Cox Point Process by Minimum Contrast

Description

Fits a log-Gaussian Cox point process model (with exponential covariance function) to a point pattern dataset by the Method of Minimum Contrast.

Usage


lgcp.estK(X, startpar=list(sigma2=1,alpha=1), lambda=NULL,
q = 1/4, p = 2, rmin = NULL, rmax = NULL)

Arguments

X 	 Data to which the model will be fitted. Either a point pattern or a summary statistic. See Details.

startpar 	 Vector of starting values for the parameters of the log-Gaussian Cox process model.

lambda 	 Optional. An estimate of the intensity of the point process.

q,p 	 Optional. Exponents for the contrast criterion.

rmin, rmax 	 Optional. The interval of r values for the contrast criterion.

Details

This algorithm fits a log-Gaussian Cox point process model to a point pattern dataset by the Method of Minimum Contrast, using the K function.

The argument X can be either

a point pattern:

An object of class "ppp" representing a point pattern dataset. The K function of the point pattern will be computed using Kest, and the method of minimum contrast will be applied to this.

a summary statistic:

An object of class "fv" containing the values of a summary statistic, computed for a point pattern dataset. The summary statistic should be the K function, and this object should have been obtained by a call to Kest or one of its relatives.

The algorithm fits a log-Gaussian Cox point process (LGCP) model to X, by finding the parameters of the LGCP model which give the closest match between the theoretical K function of the LGCP model and the observed K function. For a more detailed explanation of the Method of Minimum Contrast, see mincontrast.

The model fitted is a stationary, isotropic log-Gaussian Cox process with exponential covariance (Moller and Waagepetersen, 2003, pp. 72-76). To define this process we start with a stationary Gaussian random field Z in the two-dimensional plane, with constant mean mu and covariance function

c(r) = sigma^2 * exp(-r/alpha)

where sigma^2 and alpha are parameters. Given Z, we generate a Poisson point process Y with intensity function lambda(u) = exp(Z(u)) at location u. Then Y is a log-Gaussian Cox process.

The theoretical K-function of the LGCP is

K(r) = integral from 0 to r of (2 * pi * s * exp(sigma^2 * exp(-s/alpha))) ds.

The theoretical intensity of the LGCP is

lambda= exp(mu + sigma^2/2).

In this algorithm, the Method of Minimum Contrast is first used to find optimal values of the parameters sigma^2 and alpha^2. Then the remaining parameter mu is inferred from the estimated intensity lambda.

If the argument lambda is provided, then this is used as the value of lambda. Otherwise, if X is a point pattern, then lambda will be estimated from X. If X is a summary statistic and lambda is missing, then the intensity lambda cannot be estimated, and the parameter mu will be returned as NA.

The remaining arguments rmin,rmax,q,p control the method of minimum contrast; see mincontrast.

Value

An object of class "minconfit". There are methods for printing and plotting this object. It contains the following main components:

par 	 Vector of fitted parameter values.

fit 	 Function value table (object of class "fv") containing the observed values of the summary statistic (observed) and the theoretical values of the summary statistic computed from the fitted model parameters.

Author(s)

Rasmus Waagepetersen rw@math.auc.dk. Adapted for spatstat by Adrian Baddeley adrian@maths.uwa.edu.au http://www.maths.uwa.edu.au/~adrian/

References

Moller, J. and Waagepetersen, R. (2003). Statistical Inference and Simulation for Spatial Point Processes. Chapman and Hall/CRC, Boca Raton.

Waagepetersen, R. (2006). An estimation function approach to inference for inhomogeneous Neyman-Scott processes. Submitted.

See Also

lgcp.estK, mincontrast, Kest

Examples


    data(redwood)
    u <- lgcp.estK(redwood, c(sigma2=0.1, alpha=1))
    u
    plot(u)

[Package spatstat version 1.10-3 Index]