matclust.estK {spatstat}	R Documentation

Fit the Matern Cluster Point Process by Minimum Contrast

Description

Fits the Matern Cluster point process to a point pattern dataset by the Method of Minimum Contrast.

Usage


matclust.estK(X, startpar=c(kappa=1,R=1), lambda=NULL,
            q = 1/4, p = 2, rmin = NULL, rmax = NULL)

Arguments

X 	 Data to which the Matern Cluster model will be fitted. Either a point pattern or a summary statistic. See Details.

startpar 	 Vector of starting values for the parameters of the Matern Cluster process.

lambda 	 Optional. An estimate of the intensity of the point process.

q,p 	 Optional. Exponents for the contrast criterion.

rmin, rmax 	 Optional. The interval of r values for the contrast criterion.

Details

This algorithm fits the Matern Cluster point process model to a point pattern dataset by the Method of Minimum Contrast, using the K function.

The argument X can be either

a point pattern:

An object of class "ppp" representing a point pattern dataset. The K function of the point pattern will be computed using Kest, and the method of minimum contrast will be applied to this.

a summary statistic:

An object of class "fv" containing the values of a summary statistic, computed for a point pattern dataset. The summary statistic should be the K function, and this object should have been obtained by a call to Kest or one of its relatives.

The algorithm fits the Matern Cluster point process to X, by finding the parameters of the Matern Cluster model which give the closest match between the theoretical K function of the Matern Cluster process and the observed K function. For a more detailed explanation of the Method of Minimum Contrast, see mincontrast.

The Matern Cluster point process is described in Moller and Waagepetersen (2003, p. 62). It is a cluster process formed by taking a pattern of parent points, generated according to a Poisson process with intensity kappa, and around each parent point, generating a random number of offspring points, such that the number of offspring of each parent is a Poisson random variable with mean mu, and the locations of the offspring points of one parent are independent and uniformly distributed inside a circle of radius R centred on the parent point.

The theoretical K-function of the Matern Cluster process is

K(r) = pi r^2 + h(r/(2*R))/kappa

where

h(z) = 2 + (1/pi) * ((8 * z^2 - 4) * arccos(z) - 2 * arcsin(z) + 4 * z * sqrt((1 - z^2)^3) - 6 * z * sqrt(1 - z^2))

for z <= 1, and h(z) = 1 for z > 1. The theoretical intensity of the Matern Cluster process is lambda=kappa* mu.

In this algorithm, the Method of Minimum Contrast is first used to find optimal values of the parameters kappa and R. Then the remaining parameter mu is inferred from the estimated intensity lambda.

If the argument lambda is provided, then this is used as the value of lambda. Otherwise, if X is a point pattern, then lambda will be estimated from X. If X is a summary statistic and lambda is missing, then the intensity lambda cannot be estimated, and the parameter mu will be returned as NA.

The remaining arguments rmin,rmax,q,p control the method of minimum contrast; see mincontrast.

The Matern Cluster process can be simulated, using rMatClust.

Value

An object of class "minconfit". There are methods for printing and plotting this object. It contains the following main components:

par 	 Vector of fitted parameter values.

fit 	 Function value table (object of class "fv") containing the observed values of the summary statistic (observed) and the theoretical values of the summary statistic computed from the fitted model parameters.

Author(s)

Rasmus Waagepetersen rw@math.auc.dk Adapted for spatstat by Adrian Baddeley adrian@maths.uwa.edu.au http://www.maths.uwa.edu.au/~adrian/

References

Moller, J. and Waagepetersen, R. (2003). Statistical Inference and Simulation for Spatial Point Processes. Chapman and Hall/CRC, Boca Raton.

Waagepetersen, R. (2006). An estimation function approach to inference for inhomogeneous Neyman-Scott processes. Submitted.

See Also

lgcp.estK, mincontrast, Kest, rMatClust to simulate the fitted model.

Examples


    data(redwood)
    u <- matclust.estK(redwood, c(kappa=10, R=0.1))
    u
    plot(u)

[Package spatstat version 1.10-3 Index]