ord.family {spatstat}	R Documentation

Ord Interaction Process Family

Description

An object describing the family of all Ord interaction point processes

Details

Advanced Use Only!

This structure would not normally be touched by the user. It describes the family of point process models introduced by Ord (1977).

If you need to create a specific Ord-type model for use in analysis, use the function OrdThresh or Ord.

Anyway, ord.family is an object of class "isf" containing a function ord.family$eval for evaluating the sufficient statistics of any Ord type point process model taking an exponential family form.

Author(s)

Adrian Baddeley adrian@maths.uwa.edu.au http://www.maths.uwa.edu.au/~adrian/ and Rolf Turner rolf@math.unb.ca http://www.math.unb.ca/~rolf

References

Baddeley, A. and Turner, R. (2000) Practical maximum pseudolikelihood for spatial point patterns. Australian and New Zealand Journal of Statistics 42, 283–322.

Ord, J.K. (1977) Contribution to the discussion of Ripley (1977).

Ord, J.K. (1978) How many trees in a forest? Mathematical Scientist 3, 23–33.

Ripley, B.D. (1977) Modelling spatial patterns (with discussion). Journal of the Royal Statistical Society, Series B, 39, 172 – 212.

See Also

pairwise.family, pairsat.family, Poisson, Pairwise, PairPiece, Strauss, StraussHard, Softcore, Geyer, Saturated, Ord, OrdThresh

[Package spatstat version 1.10-3 Index]