pairdist {spatstat}	R Documentation

Pairwise distances

Description

Computes the matrix of distances between all pairs of `things' in a dataset

Usage


  pairdist(X, ...)

Arguments

X 	 Object specifying the locations of a set of `things' (such as a set of points or a set of line segments).

... 	 Further arguments depending on the method.

Details

Given a dataset X and Y (representing either a point pattern or a line segment pattern) pairdist computes the distance between each pair of `things' in the dataset, and returns a matrix containing these distances.

The function pairdist is generic, with methods for point patterns (objects of class "ppp"), line segment patterns (objects of class "psp") and a default method. See the documentation for pairdist.ppp, pairdist.psp or pairdist.default for details.

Value

A square matrix whose [i,j] entry is the distance between the `things' numbered i and j.

Author(s)

Pavel Grabarnik pavel.grabar@issp.serpukhov.su and Adrian Baddeley adrian@maths.uwa.edu.au http://www.maths.uwa.edu.au/~adrian/

See Also

pairdist.ppp, pairdist.psp, pairdist.default, crossdist, nndist, Kest

[Package spatstat version 1.10-3 Index]