pairdist.ppp {spatstat}	R Documentation

Pairwise distances

Description

Computes the matrix of distances between all pairs of points in a point pattern.

Usage


  ## S3 method for class 'ppp':
  pairdist(X, ..., periodic=FALSE, method="C")

Arguments

X 	 A point pattern (object of class "ppp").

... 	 Ignored.

periodic 	 Logical. Specifies whether to apply a periodic edge correction.

method 	 String specifying which method of calculation to use. Values are "C" and "interpreted". Usually not specified.

Details

This is a method for the generic function pairdist.

Given a point pattern X (an object of class "ppp"), this function computes the Euclidean distances between all pairs of points in X, and returns the matrix of distances.

Alternatively if periodic=TRUE and the window containing X is a rectangle, then the distances will be computed in the `periodic' sense (also known as `torus' distance): opposite edges of the rectangle are regarded as equivalent. This is meaningless if the window is not a rectangle.

The argument method is not normally used. It is retained only for checking the validity of the software. If method = "interpreted" then the distances are computed using interpreted R code only. If method="C" (the default) then C code is used. The C code is somewhat faster.

Value

A square matrix whose [i,j] entry is the distance between the points numbered i and j.

Author(s)

Pavel Grabarnik pavel.grabar@issp.serpukhov.su and Adrian Baddeley adrian@maths.uwa.edu.au http://www.maths.uwa.edu.au/~adrian/

See Also

pairdist, pairdist.default, pairdist.psp, crossdist, nndist, Kest

Examples


   data(cells)
   d <- pairdist(cells)
   d <- pairdist(cells, periodic=TRUE)

[Package spatstat version 1.10-3 Index]