pairwise.family {spatstat}	R Documentation

Pairwise Interaction Process Family

Description

An object describing the family of all pairwise interaction point processes

Details

Advanced Use Only!

This structure would not normally be touched by the user. It describes the pairwise interaction family of point process models.

If you need to create a specific pairwise interaction model for use in modelling, use the function Pairwise or one of the existing functions (listed below, starting with a capital letter).

Anyway, pairwise.family is an object of class "isf" containing a function pairwise$eval for evaluating the sufficient statistics of any pairwise interaction point process model taking an exponential family form.

Author(s)

Adrian Baddeley adrian@maths.uwa.edu.au http://www.maths.uwa.edu.au/~adrian/ and Rolf Turner rolf@math.unb.ca http://www.math.unb.ca/~rolf

See Also

pairsat.family, ord.family, Poisson, Pairwise, PairPiece, Strauss, StraussHard, Softcore, Geyer, Saturated, Ord, OrdThresh

[Package spatstat version 1.10-3 Index]