plot.fasp {spatstat}	R Documentation

Plot a Function Array

Description

Plots an array of summary functions, usually associated with a point pattern, stored in an object of class "fasp". A method for plot.

Usage


   ## S3 method for class 'fasp':
   plot(x,formule=NULL, ...,
                        subset=NULL, title=NULL, samex=TRUE)

Arguments

x 	 An object of class "fasp" representing a function array.

formule 	 A formula or list of formulae indicating what variables are to be plotted against what variable. Each formula is either an R language formula object, or a string that can be parsed as a formula. If formule is a list, its k-th component should be applicable to the (i,j)-th plot where x$which[i,j]=k. If the formula is left as NULL, then plot.fasp attempts to use the component default.formula of x. If that component is NULL as well, it gives up.

subset 	 A logical vector, or a vector of indices, or an expression or a character string, or a list of such, indicating a subset of the data to be included in each plot. If subset is a list, its k-th component should be applicable to the (i,j)-th plot where x$which[i,j]=k.

title 	 Overall title for the plot.

samex 	 Logical flag indicating whether all individual plots should have the same x axis limits. This makes it easier to compare the plots. It can only be set to FALSE if you are using the default plot style (i.e. only when formule is missing).

... 	 Arguments passed to plot.fv to control the individual plot panels.

Details

An object of class "fasp" represents an array of summary functions, usually associated with a point pattern. See fasp.object for details. Such an object might be created, for example, by alltypes or allstats.

The function plot.fasp is a method for plot. It calls plot.fv to plot the individual panels.

For information about the interpretation of the arguments formule and subset, see plot.fv.

Arguments that are often passed through ... include col to control the colours of the different lines in a panel, and lty and lwd to control the line type and line width of the different lines in a panel. See plot.fv.

The argument title, if present, will determine the overall title of the plot. If it is absent, it defaults to x$title. Titles for the individual plot panels will be taken from x$titles.

Value

None.

Warnings

(Each component of) the subset argument may be a logical vector (of the same length as the vectors of data which are extracted from x), or a vector of indices, or an expression such as expression(r<=0.2), or a text string, such as "r<=0.2".

Attempting a syntax such as subset = r<=0.2 (without wrapping r<=0.2 either in quote marks or in expression()) will cause this function to fall over.

Variables referred to in any formula must exist in the data frames stored in x. What the names of these variables are will of course depend upon the nature of x.

Author(s)

Adrian Baddeley adrian@maths.uwa.edu.au http://www.maths.uwa.edu.au/~adrian/ and Rolf Turner rolf@math.unb.ca http://www.math.unb.ca/~rolf

See Also

alltypes, allstats, plot.fv, fasp.object

Examples


   ## Not run: 
   # Bramble Canes data.
   data(bramblecanes)

   X.G <- alltypes(bramblecanes,type="G",dataname="Bramblecanes",verb=TRUE)
   plot(X.G)
   plot(X.G,subset="r<=0.2")
   plot(X.G,formule=cbind(asin(sqrt(km)),
                    asin(sqrt(theo))) ~ asin(sqrt(theo)))
       plot(X.G,fo=cbind(km-theo,0)~r,"r<=0.2")

   # Swedish pines.
   data(swedishpines)
   X <- allstats(swedishpines,dataname="Swedish Pines")
   plot(X,subset=list("r<=20","r<=20","r<=20","r<=50"))

   # Simulated data.
   pp <- runifpoint(350, owin(c(0,1),c(0,1)))
   pp$marks <- factor(c(rep(1,50),rep(2,100),rep(3,200)))
   X.K <- alltypes(pp,type="K",verb=TRUE,dataname="Fake Data")
   plot(X.K,fo=cbind(border,theo)~theo,"theo<=0.75")
   ## End(Not run)

[Package spatstat version 1.10-3 Index]