plot.listof {spatstat}	R Documentation

Plot a List of Things

Description

Plots a list of things

Usage


  ## S3 method for class 'listof':
  plot(x, ..., main, arrange=TRUE)

Arguments

x 	 An object of the class "listof". Essentially a list of objects.

... 	 Arguments passed to plot when generating each plot panel.

main 	 Overall heading for the plot.

arrange 	 Logical flag indicating whether to plot the objects side-by-side. Set this to FALSE if you want to control the arrangement of the plots yourself.

Details

This is the plot method for the class "listof".

An object of class "listof" (defined in the base R package) represents a list of objects, all belonging to a common class. The base R package defines a method for printing these objects, print.listof, but does not define a method for plot. So here we have provided a method for plot.

In the spatstat package, the function density.splitppp produces an object of class "listof", essentially a list of pixel images. These images can be plotted in a nice arrangement using plot.listof. See the Example.

Value

Null.

Author(s)

Adrian Baddeley adrian@maths.uwa.edu.au http://www.maths.uwa.edu.au/~adrian/ and Rolf Turner rolf@math.unb.ca http://www.math.unb.ca/~rolf

See Also

print.listof, density.splitppp

Examples


# Multitype point pattern
 data(amacrine)
 plot(density(split(amacrine)))

[Package spatstat version 1.10-3 Index]