plot.plotppm {spatstat}	R Documentation

Plot a plotppm Object Created by plot.ppm

Description

The function plot.ppm produces objects which specify plots of fitted point process models. The function plot.plotppm carries out the actual plotting of these objects.

Usage


  ## S3 method for class 'plotppm':
  plot(x, data = NULL, trend = TRUE, cif = TRUE, pause = interactive(),
             how = c("persp", "image", "contour"), ...)

Arguments

x 	 An object of class plotppm produced by plot.ppm()

data 	 The point pattern (an object of class ppp) to which the point process model was fitted (by code{ppm()}).

trend 	 Logical scalar; should the trend component of the fitted model be plotted?

cif 	 Logical scalar; should the complete conditional intensity of the fitted model be plotted?

pause 	 Logical scalar indicating whether to pause with a prompt after each plot. Set pause=FALSE if plotting to a file.

how 	 Character string or character vector indicating the style or styles of plots to be performed.

... 	 Extra arguments to the plotting functions persp, image and contour.

Details

If argument data is supplied then the point pattern will be superimposed on the image and contour plots.

Sometimes a fitted model does not have a trend component, or the trend component may constitute all of the conditional intensity (if the model is Poisson). In such cases the object x will not contain a trend component, or will contain only a trend component. This will also be the case if one of the arguments trend and cif was set equal to FALSE in the call to plot.ppm() which produced x. If this is so then only the item which is present will be plotted. Explicitly setting trend=TRUE, or cif=TRUE, respectively, will then give an error.

Value

None.

Warning

Arguments which are passed to persp, image, and contour via the ... argument get passed to any of the other functions listed in the how argument, and won't be recognized by them. This leads to a lot of annoying but harmless warning messages. Arguments to persp may be supplied via spatstat.options() which alleviates the warning messages in this instance.

Author(s)

Adrian Baddeley adrian@maths.uwa.edu.au http://www.maths.uwa.edu.au/~adrian/ and Rolf Turner rolf@math.unb.ca http://www.math.unb.ca/~rolf

See Also

plot.ppm()

Examples


 ## Not run: 
 data(cells)
 Q <- quadscheme(cells)
 m <- ppm(Q, ~1, Strauss(0.05))
 mpic <- plot(m)
 # Perspective plot only, with altered parameters:
  plot(mpic,how="persp", theta=-30,phi=40,d=4)
 # All plots, with altered parameters for perspective plot:
 spatstat.options(par.persp=list(theta=-30,phi=40,d=4))
 plot(mpic)
 
## End(Not run)

[Package spatstat version 1.10-3 Index]