plot.quad {spatstat}	R Documentation

plot a Spatial Quadrature Scheme

Description

Plot a two-dimensional spatial quadrature scheme.

Usage


  ## S3 method for class 'quad':
  plot(x, ..., main=deparse(substitute(x)), dum=list())

Arguments

x 	 The spatial quadrature scheme to be plotted. An object of class "quad".

... 	 extra arguments controlling the plotting of the data points of the quadrature scheme.

dum 	 list of extra arguments controlling the plotting of the dummy points of the quadrature scheme. See below.

main 	 text to be displayed as a title above the plot.

Details

This is the plot method for quadrature schemes (objects of class "quad", see quad.object).

First the data points of the quadrature scheme are plotted (in their observation window) using plot.ppp with any arguments specified in ...

Then the dummy points of the quadrature scheme are plotted using plot.ppp with any arguments specified in dum.

By default the dummy points are superimposed onto the plot of data points. This can be overridden by including the argument add=FALSE in the list dum as shown in the examples. In this case the data and dummy point patterns are plotted separately.

See par and plot.ppp for other possible arguments controlling the plots.

Value

NULL.

Author(s)

Adrian Baddeley adrian@maths.uwa.edu.au http://www.maths.uwa.edu.au/~adrian/ and Rolf Turner rolf@math.unb.ca http://www.math.unb.ca/~rolf

See Also

quad.object, plot.ppp, par

Examples


   data(nztrees)
   Q <- quadscheme(nztrees)

   plot(Q, main="NZ trees: quadrature scheme")

   oldpar <- par(mfrow=c(2,1))
   plot(Q, main="NZ trees", dum=list(add=FALSE))
   par(oldpar)

[Package spatstat version 1.10-3 Index]