plot.splitppp {spatstat}	R Documentation

Plot a List of Point Patterns

Description

Plots a list of point patterns.

Usage


  ## S3 method for class 'splitppp':
  plot(x, ..., main, arrange=TRUE)

Arguments

x 	 A named list of point patterns, typically obtained from split.ppp.

... 	 Arguments passed to plot.ppp which control the appearance of each plot panel.

main 	 Overall heading for the plot.

arrange 	 Logical flag indicating whether to plot the point patterns side-by-side. Set this to FALSE if you want to control the arrangement of the plots yourself.

Details

This is the plot method for the class "splitppp". It is typically used to plot the result of the function split.ppp but it may also be used to plot any list of point patterns created by the user.

The argument x should be a named list of point patterns (objects of class "ppp", see ppp.object). Each of these point patterns will be plotted in turn using plot.ppp.

Value

Null.

Author(s)

Adrian Baddeley adrian@maths.uwa.edu.au http://www.maths.uwa.edu.au/~adrian/ and Rolf Turner rolf@math.unb.ca http://www.math.unb.ca/~rolf

See Also

split.ppp, plot.ppp, ppp.object

Examples


# Multitype point pattern
 data(amacrine)
 plot(split(amacrine))

[Package spatstat version 1.10-3 Index]