ppm.object {spatstat}	R Documentation

Class of Fitted Point Process Models

Description

A class ppm to represent a fitted stochastic model for a point process. The output of ppm.

Details

An object of class ppm represents a stochastic point process model that has been fitted to a point pattern dataset. Typically it is the output of the model fitter, ppm.

There are methods print.ppm, plot.ppm, predict.ppm, fitted.ppm and coef.ppm for the generic functions print, plot, predict, fitted and coef respectively.

See also (for example) Strauss to understand how to specify a point process model with unknown parameters.

Information about the data (to which the model was fitted) can be extracted using data.ppm, dummy.ppm and quad.ppm.

If you really need to get at the internals, a ppm object contains at least the following entries:

coef 	 the fitted regular parameters (as returned by glm)

trend 	 the trend formula or NULL

interaction 	 the point process interaction family (an object of class "interact") or NULL

Q 	 the quadrature scheme used

maxlogpl 	 the maximised value of log pseudolikelihood

correction 	 name of edge correction method used

See ppm for explanation of these concepts. The irregular parameters (e.g. the interaction radius of the Strauss process) are encoded in the interaction entry. However see the Warnings.

Warnings

The internal representation may change in the next few releases of the spatstat package.

Author(s)

Adrian Baddeley adrian@maths.uwa.edu.au http://www.maths.uwa.edu.au/~adrian/ and Rolf Turner rolf@math.unb.ca http://www.math.unb.ca/~rolf

See Also

ppm, coef.ppm, fitted.ppm, print.ppm, predict.ppm, plot.ppm.

Examples


  data(cells)
  fit <- ppm(cells, ~ x, Strauss(0.1), correction="periodic")
  fit
  coef(fit)
  ## Not run: 
  pred <- predict(fit)
  
## End(Not run)
  pred <- predict(fit, ngrid=20, type="trend")
  ## Not run: 
  plot(fit)
  
## End(Not run)  

[Package spatstat version 1.10-3 Index]