print.im {spatstat}	R Documentation

Print Brief Details of an Image

Description

Prints a very brief description of a pixel image object.

Usage


  ## S3 method for class 'im':
  print(x, ...)

Arguments

x 	 Pixel image (object of class "im").

... 	 Ignored.

Details

A very brief description of the pixel image x is printed.

This is a method for the generic function print.

Author(s)

Adrian Baddeley adrian@maths.uwa.edu.au http://www.maths.uwa.edu.au/~adrian/ and Rolf Turner rolf@math.unb.ca http://www.math.unb.ca/~rolf

See Also

print, im.object, summary.im

Examples


  data(letterR)
  U <- as.im(letterR)
  U

[Package spatstat version 1.10-3 Index]