print.owin {spatstat}	R Documentation

Print Brief Details of a Spatial Window

Description

Prints a very brief description of a window object.

Usage


  ## S3 method for class 'owin':
  print(x, ...)

Arguments

x 	 Window (object of class "owin").

... 	 Ignored.

Details

A very brief description of the window x is printed.

This is a method for the generic function print.

Author(s)

Adrian Baddeley adrian@maths.uwa.edu.au http://www.maths.uwa.edu.au/~adrian/ and Rolf Turner rolf@math.unb.ca http://www.math.unb.ca/~rolf

See Also

print, print.ppp, summary.owin

Examples


  owin()  # the unit square

  data(demopat)
  W <- demopat$window  
  W                    # just says it is polygonal
  as.mask(W)           # just says it is a binary image 


[Package spatstat version 1.10-3 Index]