print.ppm {spatstat}	R Documentation

Print a Fitted Point Process Model

Description

Default print method for a fitted point process model.

Usage


 ## S3 method for class 'ppm':
 print(x,...)

Arguments

x 	 A fitted point process model, typically obtained from the model-fittingg algorithm ppm. An object of class "ppm".

... 	 Ignored.

Details

This is the print method for the class "ppm". It prints information about the fitted model in a sensible format.

See ppm.object for details of the class "ppm".

Value

none.

Author(s)

Adrian Baddeley adrian@maths.uwa.edu.au http://www.maths.uwa.edu.au/~adrian/ and Rolf Turner rolf@math.unb.ca http://www.math.unb.ca/~rolf

See Also

ppm, ppm.object, plot.ppm, predict.ppm

Examples


 ## Not run: 
 data(cells)
 Q <- quadscheme(cells)
 m <- ppm(Q, ~1, Strauss(0.05))
 m
 
## End(Not run)

[Package spatstat version 1.10-3 Index]