print.ppp {spatstat}	R Documentation

Print Brief Details of a Point Pattern Dataset

Description

Prints a very brief description of a point pattern dataset.

Usage


  ## S3 method for class 'ppp':
  print(x, ...)

Arguments

x 	 Point pattern (object of class "ppp").

... 	 Ignored.

Details

A very brief description of the point pattern x is printed.

This is a method for the generic function print.

Author(s)

Adrian Baddeley adrian@maths.uwa.edu.au http://www.maths.uwa.edu.au/~adrian/ and Rolf Turner rolf@math.unb.ca http://www.math.unb.ca/~rolf

See Also

print, print.owin, summary.ppp

Examples


  data(cells)      # plain vanilla point pattern
  cells

  data(lansing)    # multitype point pattern
  lansing          
  
  data(longleaf)    # numeric marks
  longleaf          

  data(demopat)     # weird polygonal window
  demopat

[Package spatstat version 1.10-3 Index]