quadrat.test {spatstat}	R Documentation

Chi-Squared Dispersion Test for Spatial Point Pattern Based on Quadrat Counts

Description

Performs a chi-squared test of complete spatial randomness for a given point pattern, based on quadrat counts. Alternatively performs a chi-squared goodness-of-fit test of a fitted inhomogeneous Poisson model.

Usage


quadrat.test(X, nx = 5, ny = nx, xbreaks = NULL, ybreaks = NULL, fit)

Arguments

X 	 A point pattern (object of class "ppp") to be subjected to the goodness-of-fit test. Alternatively a fitted point process model (object of class "ppm") to be tested.

nx,ny 	 Numbers of quadrats in the x and y directions. Incompatible with xbreaks and ybreaks.

xbreaks 	 Optional. Numeric vector giving the x coordinates of the boundaries of the quadrats. Incompatible with nx.

ybreaks 	 Optional. Numeric vector giving the y coordinates of the boundaries of the quadrats. Incompatible with ny.

fit 	 Optional. A fitted point process model (object of class "ppm"). The point pattern X will be subjected to a test of goodness-of-fit to the model fit.

Details

This function performs a chi^2 test of goodness-of-fit to the Poisson point process (including ‘Complete Spatial Randomness’ but also inhomogeneous Poisson processes), based on quadrat counts.

If X is a point pattern, it is taken as the data point pattern for the test. If X is a fitted point process model, then the data to which this model was fitted are extracted from the model object, and are treated as the data point pattern for the test.

The window of observation is divided into rectangular tiles and the number of data points in each tile is counted, as described in quadratcount.

If fit is absent, then we test the null hypothesis that the data pattern is a realisation of Complete Spatial Randomness (the uniform Poisson point process) by applying the chi^2 test of goodness-of-fit to the quadrat counts.

If fit is present, then it should be a point process model (object of class "ppm") and it should be a Poisson point process. Then we test the null hypothesis that the data pattern is a realisation of the (inhomogeneous) Poisson point process specified by fit. Again this is a chi^2 test of goodness-of-fit to the quadrat counts.

The return value is an object of class "htest". Printing the object gives comprehensible output about the outcome of the test. The return value also belongs to the special class "quadrat.test". Plotting the object will display the quadrats, annotated by their observed and expected counts and the Pearson residuals. See the examples.

To test the Poisson point process against a specific alternative point process model, use anova.ppm.

Value

An object of class "htest". See chisq.test for explanation.
The return value is also an object of the special class "quadrat.test", and there is a plot method for this class. See the examples.

Author(s)

Adrian Baddeley adrian@maths.uwa.edu.au http://www.maths.uwa.edu.au/~adrian/ and Rolf Turner rolf@math.unb.ca http://www.math.unb.ca/~rolf

See Also

quadratcount, chisq.test

Examples


  data(simdat)
  quadrat.test(simdat)
  quadrat.test(simdat, 4)

  # fitted model: inhomogeneous Poisson
  fitx <- ppm(simdat, ~x, Poisson())
  # equivalent:
  quadrat.test(simdat, fit=fitx)
  quadrat.test(fitx)

  te <- quadrat.test(simdat, 4)
  residuals(te)  # Pearson residuals

  plot(te)

  plot(simdat, pch="+", col="green", cex=1.2, lwd=2)
  plot(te, add=TRUE, col="red", cex=1.5, lty=2, lwd=3)

  sublab <- eval(substitute(expression(p[chi^2]==z),
                       list(z=signif(te$p.value,3))))
  title(sub=sublab, cex.sub=3)


[Package spatstat version 1.10-3 Index]