quadratcount {spatstat}	R Documentation

Quadrat counting for a point pattern

Description

Divides window into quadrats and counts the numbers of points in each quadrat.

Usage


  quadratcount(X, nx=5, ny=nx, xbreaks, ybreaks)

Arguments

X 	 A point pattern (object of class "ppp").

nx,ny 	 Numbers of quadrats in the x and y directions. Incompatible with xbreaks and ybreaks.

xbreaks 	 Numeric vector giving the x coordinates of the boundaries of the quadrats. Incompatible with nx.

ybreaks 	 Numeric vector giving the y coordinates of the boundaries of the quadrats. Incompatible with ny.

Details

Quadrat counting is an elementary technique for analysing spatial point patterns. See Diggle (2003).

The window containing the point pattern X is divided into an nx * ny grid of rectangular tiles or `quadrats'. The number of points of X falling in each quadrat is counted. These numbers are returned as a contingency table.

If xbreaks is given, it should be a numeric vector giving the x coordinates of the quadrat boundaries. If it is not given, it defaults to a sequence of nx+1 values equally spaced over the range of x coordinates in the window X$window.

Similarly if ybreaks is given, it should be a numeric vector giving the y coordinates of the quadrat boundaries. It defaults to a vector of ny+1 values equally spaced over the range of y coordinates in the window. The lengths of xbreaks and ybreaks may be different.

The algorithm counts the number of points of X falling in each quadrat, and returns these counts as a contingency table. The [i,j] entry in the contingency table is the point count for the quadrat with coordinates (xbreaks[i],xbreaks[i+1]) by (ybreaks[i], ybreaks[i+1]).

The return value is a table which can be printed neatly. The return value is also a member of the special class "quadratcount". Plotting the object will display the quadrats, annotated by their counts. See the examples.

To perform a chi-squared test based on the quadrat counts, use quadrat.test.

Value

A contingency table containing the number of points in each quadrat.
The table is also an object of the special class "quadratcount" and there is a plot method for this class.

Author(s)

Adrian Baddeley adrian@maths.uwa.edu.au http://www.maths.uwa.edu.au/~adrian/ and Rolf Turner rolf@math.unb.ca http://www.math.unb.ca/~rolf

References

Diggle, P.J. Statistical analysis of spatial point patterns. Academic Press, 2003.

Stoyan, D. and Stoyan, H. (1994) Fractals, random shapes and point fields: methods of geometrical statistics. John Wiley and Sons.

See Also

quadrat.test

Examples


 X <- runifpoint(50)
 quadratcount(X)
 quadratcount(X, 4, 5)
 quadratcount(X, xbreaks=c(0, 0.3, 1), ybreaks=c(0, 0.4, 0.8, 1))
 qX <-  quadratcount(X, 4, 5)

 # plotting:
 plot(X, pch="+")
 plot(qX, add=TRUE, col="red", cex=1.5, lty=2)

[Package spatstat version 1.10-3 Index]