rMatClust {spatstat}	R Documentation

Simulate Matern Cluster Process

Description

Generate a random point pattern, a simulated realisation of the Mat'ern Cluster Process.

Usage


 rMatClust(kappa, r, mu, win = owin(c(0,1),c(0,1)))

Arguments

kappa 	 Intensity of the Poisson process of cluster centres. A single positive number, a function, or a pixel image.

r 	 Radius parameter of the clusters.

mu 	 Mean number of points per cluster.

win 	 Window in which to simulate the pattern. An object of class "owin" or something acceptable to as.owin.

Details

This algorithm generates a realisation of Mat'ern's cluster process inside the window win. The process is constructed by first generating a Poisson point process of ``parent'' points with intensity kappa. Then each parent point is replaced by a random cluster of points, the number of points in each cluster being random with a Poisson (mu) distribution, and the points being placed independently and uniformly inside a disc of radius r centred on the parent point.

In this implementation, parent points are not restricted to lie in the window; the parent process is effectively the uniform Poisson process on the infinite plane.

Value

The simulated point pattern (an object of class "ppp").
Additionally, some intermediate results of the simulation are returned as attributes of this point pattern. See rNeymanScott.

Author(s)

Adrian Baddeley adrian@maths.uwa.edu.au http://www.maths.uwa.edu.au/~adrian/ and Rolf Turner rolf@math.unb.ca http://www.math.unb.ca/~rolf

See Also

rpoispp, rThomas, rNeymanScott

Examples


 pp <- rMatClust(10, 0.05, 4)

[Package spatstat version 1.10-3 Index]