rMaternI {spatstat}	R Documentation

Simulate Matern Model I

Description

Generate a random point pattern, a simulated realisation of the Mat'ern Model I inhibition process model.

Usage


 rMaternI(kappa, r, win = owin(c(0,1),c(0,1)))

Arguments

kappa 	 Intensity of the Poisson process of proposal points. A single positive number.

r 	 Inhibition distance.

win 	 Window in which to simulate the pattern. An object of class "owin" or something acceptable to as.owin.

Details

This algorithm generates a realisation of Mat'ern's Model I inhibition process inside the window win.

The process is constructed by first generating a uniform Poisson point process of ``proposal'' points with intensity kappa inside the window. A proposal point is deleted if it lies within r units' distance of another proposal point. Otherwise it is retained. The retained points constitute Mat'ern's Model I.

Value

The simulated point pattern (an object of class "ppp").

Author(s)

Adrian Baddeley adrian@maths.uwa.edu.au http://www.maths.uwa.edu.au/~adrian/ and Rolf Turner rolf@math.unb.ca http://www.math.unb.ca/~rolf

See Also

rpoispp, rMatClust

Examples


 pp <- rMaternI(20, 0.05)

[Package spatstat version 1.10-3 Index]