rMaternII {spatstat}	R Documentation

Simulate Matern Model II

Description

Generate a random point pattern, a simulated realisation of the Mat'ern Model II inhibition process.

Usage


 rMaternII(kappa, r, win = owin(c(0,1),c(0,1)))

Arguments

kappa 	 Intensity of the Poisson process of proposal points. A single positive number.

r 	 Inhibition distance.

win 	 Window in which to simulate the pattern. An object of class "owin" or something acceptable to as.owin.

Details

This algorithm generates a realisation of Mat'ern's Model II inhibition process inside the window win.

The process is constructed by first generating a uniform Poisson point process of ``proposal'' points with intensity kappa inside the window. Then each proposal point is marked by an ``arrival time'', a number uniformly distributed in [0,1] independently of other variables.

A proposal point is deleted if it lies within r units' distance of another proposal point that has an earlier arrival time. Otherwise it is retained. The retained points constitute Mat'ern's Model II.

The difference between Mat'ern's Model I and II is the italicised statement above. Model II has a higher intensity for the same parameter values.

Value

The simulated point pattern (an object of class "ppp").

Author(s)

Adrian Baddeley adrian@maths.uwa.edu.au http://www.maths.uwa.edu.au/~adrian/ and Rolf Turner rolf@math.unb.ca http://www.math.unb.ca/~rolf

See Also

rpoispp, rMatClust, rMaternI

Examples


 pp <- rMaternII(20, 0.05)

[Package spatstat version 1.10-3 Index]