rSSI {spatstat}	R Documentation

Simulate Simple Sequential Inhibition

Description

Generate a random point pattern, a realisation of the Simple Sequential Inhibition (SSI) process.

Usage


 rSSI(r, n, win = owin(c(0,1),c(0,1)), giveup = 1000)

Arguments

r 	 Inhibition distance.

n 	 Number of points to generate.

win 	 Window in which to simulate the pattern. An object of class "owin" or something acceptable to as.owin.

giveup 	 Number of rejected proposals after which the algorithm should terminate.

Details

This algorithm generates a realisation of the Simple Sequential Inhibition point process inside the window win.

Starting with an empty window, the algorithm adds points one-by-one. Each new point is generated uniformly in the window and independently of preceding points. If the new point lies closer than r units from an existing point, then it is rejected and another random point is generated.

The algorithm terminates either when the desired number n of points is reached, or when the current point configuration has not changed for giveup iterations.

Value

The simulated point pattern (an object of class "ppp").

Author(s)

Adrian Baddeley adrian@maths.uwa.edu.au http://www.maths.uwa.edu.au/~adrian/ and Rolf Turner rolf@math.unb.ca http://www.math.unb.ca/~rolf

See Also

rpoispp, rMaternI, rMaternII

Examples


 pp <- rSSI(0.05, 200)
 ## Not run: plot(pp)

[Package spatstat version 1.10-3 Index]