rThomas {spatstat}	R Documentation

Simulate Thomas Process

Description

Generate a random point pattern, a realisation of the Thomas cluster process.

Usage


 rThomas(kappa, sigma, mu, win = owin(c(0,1),c(0,1)))

Arguments

kappa 	 Intensity of the Poisson process of cluster centres. A single positive number.

sigma 	 Standard deviation of displacement of a point from its cluster centre.

mu 	 Expected number of points per cluster.

win 	 Window in which to simulate the pattern. An object of class "owin" or something acceptable to as.owin.

Details

This algorithm generates a realisation of the Thomas process, a special case of the Neyman-Scott process.

The algorithm generates a uniform Poisson point process of ``parent'' points with intensity kappa. Then each parent point is replaced by a random cluster of points, the number of points per cluster being Poisson (mu) distributed, and their positions being isotropic Gaussian displacements from the cluster parent location.

Value

The simulated point pattern (an object of class "ppp").
Additionally, some intermediate results of the simulation are returned as attributes of this point pattern. See rNeymanScott.

Author(s)

Adrian Baddeley adrian@maths.uwa.edu.au http://www.maths.uwa.edu.au/~adrian/ and Rolf Turner rolf@math.unb.ca http://www.math.unb.ca/~rolf

See Also

rpoispp, rMatClust, rNeymanScott

Examples


  X <- rThomas(10, 0.2, 5)

[Package spatstat version 1.10-3 Index]