rcell {spatstat}	R Documentation

Simulate Baddeley-Silverman Cell Process

Description

Generates a random point pattern, a simulated realisation of the Baddeley-Silverman cell process model.

Usage


 rcell(win=square(1), nx, ny=nx, dx=NULL, dy=NULL)

Arguments

win 	 A window. An object of class owin, or data in any format acceptable to as.owin().

dx 	 Width of the cells. Incompatible with nx.

dy 	 Height of the cells. Incompatible with ny.

nx 	 Number of columns of cells in the window. Incompatible with dx.

ny 	 Number of rows of cells in the window. Incompatible with dy.

Details

This function generates a simulated realisation of the “cell process” (Baddeley and Silverman, 1984), a random point process with the same second-order properties as the uniform Poisson process. In particular, the K function of this process is identical to the K function of the uniform Poisson process (aka Complete Spatial Randomness). The same holds for the pair correlation function and all other second-order properties. The cell process is a counterexample to the claim that the K function completely characterises a point pattern.

A cell process is generated by dividing space into equal rectangular tiles. In each tile, a random number N of points is placed, where N takes the values 0, 1 and 10 with probabilities 1/10, 8/9 and 1/90 respectively. The points within a tile are independent and uniformly distributed in that tile, and the numbers of points in different tiles are independent random integers.

In the function rcell the tile dimensions are determined by the quantities dx, dy if they are present. If they are absent, then the grid spacing is determined so that there will be nx columns and ny rows of tiles in the bounding rectangle of win. The cell process is then generated in these tiles.

Some of the resulting random points may lie outside the window win: if they do, they are deleted. The result is a point pattern inside the window win.

Value

A point pattern (object of class "ppp").

Author(s)

Adrian Baddeley adrian@maths.uwa.edu.au http://www.maths.uwa.edu.au/~adrian/ and Rolf Turner rolf@math.unb.ca http://www.math.unb.ca/~rolf

References

Baddeley, A.J. and Silverman, B.W. (1984) A cautionary example on the use of second-order methods for analyzing point patterns. Biometrics 40, 1089-1094.

See Also

rstrat, rsyst, runifpoint, Kest

Examples


  X <- rcell(nx=15)
  plot(X)
  plot(Kest(X))

[Package spatstat version 1.10-3 Index]