rescale {spatstat}	R Documentation

Convert dataset to another unit of length

Description

Converts between different units of length in a spatial dataset, such as a point pattern or a window.

Usage


  rescale(X, s)

Arguments

X 	 Any suitable dataset representing a two-dimensional object, such as a point pattern (object of class "ppp"), or a window (object of class "owin").

s 	 Conversion factor: the new units are s times the old units.

Details

This is generic. Methods are provided for point patterns (rescale.ppp) and windows (rescale.owin).

The spatial coordinates in the dataset X will be re-expressed in terms of a new unit of length that is s times the current unit of length given in X.

Value

Another object of the same type, representing the same data, but expressed in the new units.

Note

The result of this operation is equivalent to the original dataset. If you want to actually change the coordinates by a linear transformation, producing a dataset that is not equivalent to the original one, use affine.

Author(s)

Adrian Baddeley adrian@maths.uwa.edu.au http://www.maths.uwa.edu.au/~adrian/ and Rolf Turner rolf@math.unb.ca http://www.math.unb.ca/~rolf

See Also

units, rescale.ppp, rescale.owin, affine, rotate, shift

[Package spatstat version 1.10-3 Index]