rescale.owin {spatstat}	R Documentation

Convert Window to Another Unit of Length

Description

Converts a window to another unit of length.

Usage


  ## S3 method for class 'owin':
  rescale(X, s)

Arguments

X 	 Window (object of class "owin").

s 	 Conversion factor: the new units are s times the old units.

Details

This is a method for the generic function rescale.

The spatial coordinates in the window X (and its window) will be re-expressed in terms of a new unit of length that is s times the current unit of length given in X. (Thus, the coordinate values are divided by s, while the unit value is multiplied by s).

The result is a window representing the same region of space, but re-expressed in a different unit.

Value

Another window object (of class "owin") representing the same window, but expressed in the new units.

Note

The result of this operation is equivalent to the original window. If you want to actually change the coordinates by a linear transformation, producing a window that is larger or smaller than the original one, use affine.

Author(s)

Adrian Baddeley adrian@maths.uwa.edu.au http://www.maths.uwa.edu.au/~adrian/ and Rolf Turner rolf@math.unb.ca http://www.math.unb.ca/~rolf

See Also

units, rescale, rescale.owin, affine, rotate, shift

Examples


  data(swedishpines)
  W <- swedishpines$window
  W
# coordinates are in decimetres (0.1 metre)
# convert to metres
  rescale(W, 10)

[Package spatstat version 1.10-3 Index]