rescale.ppp {spatstat}	R Documentation

Convert Point Pattern to Another Unit of Length

Description

Converts a point pattern dataset to another unit of length.

Usage


 ## S3 method for class 'ppp':
 rescale(X, s)

Arguments

X 	 Point pattern (object of class "ppp").

s 	 Conversion factor: the new units are s times the old units.

Details

This is a method for the generic function rescale.

The spatial coordinates in the point pattern X (and its window) will be re-expressed in terms of a new unit of length that is s times the current unit of length given in X. (Thus, the coordinate values are divided by s, while the unit value is multiplied by s).

The result is a point pattern representing the same data but re-expressed in a different unit.

Mark values are unchanged.

Value

Another point pattern (of class "ppp"), representing the same data, but expressed in the new units.

Note

The result of this operation is equivalent to the original point pattern. If you want to actually change the coordinates by a linear transformation, producing a point pattern that is not equivalent to the original one, use affine.

Author(s)

Adrian Baddeley adrian@maths.uwa.edu.au http://www.maths.uwa.edu.au/~adrian/ and Rolf Turner rolf@math.unb.ca http://www.math.unb.ca/~rolf

See Also

units, rescale, rescale.owin, affine, rotate, shift

Examples


# Bramble Canes data: 1 unit = 9 metres
  data(bramblecanes)
# convert to metres
  bram <- rescale(bramblecanes, 1/9)

[Package spatstat version 1.10-3 Index]