rescale.psp {spatstat}	R Documentation

Convert Line Segment Pattern to Another Unit of Length

Description

Converts a line segment pattern dataset to another unit of length.

Usage


 ## S3 method for class 'psp':
 rescale(X, s)

Arguments

X 	 Line segment pattern (object of class "psp").

s 	 Conversion factor: the new units are s times the old units.

Details

This is a method for the generic function rescale.

The spatial coordinates in the line segment pattern X (and its window) will be re-expressed in terms of a new unit of length that is s times the current unit of length given in X. (Thus, the coordinate values are divided by s, while the unit value is multiplied by s).

The result is a line segment pattern representing the same data but re-expressed in a different unit.

Mark values are unchanged.

Value

Another line segment pattern (of class "psp"), representing the same data, but expressed in the new units.

Note

The result of this operation is equivalent to the original segment pattern. If you want to actually change the coordinates by a linear transformation, producing a segment pattern that is not equivalent to the original one, use affine.

Author(s)

Adrian Baddeley adrian@maths.uwa.edu.au http://www.maths.uwa.edu.au/~adrian/ and Rolf Turner rolf@math.unb.ca http://www.math.unb.ca/~rolf

See Also

units, affine, rotate, shift

Examples


   data(copper)
   X <- copper$lines
   X
   # data are in km
   # convert to metres
   rescale(X, 1/1000)

[Package spatstat version 1.10-3 Index]