ripras {spatstat}	R Documentation

Estimate window from points alone

Description

Given an observed pattern of points, computes the Ripley-Rasson estimate of the spatial domain from which they came.

Usage


 ripras(x, y=NULL)

Arguments

x 	 vector of x coordinates of observed points, or a 2-column matrix giving x,y coordinates, or a list with components x,y giving coordinates.

y 	 (optional) vector of y coordinates of observed points, if x is a vector.

Details

Given an observed pattern of points with coordinates given by x and y, this function computes an estimate due to Ripley and Rasson (1977) of the spatial domain from which the points came.

The points are assumed to have been generated independently and uniformly distributed inside an unknown domain D. The maximum likelihood estimate of D is the convex hull of the points. Analogously to the problems of estimating the endpoint of a uniform distribution, the MLE is not optimal. Ripley and Rasson's estimator is a rescaled copy of the convex hull, centred at the centroid of the convex hull. The scaling factor is 1/sqrt{1 - frac m n} where n is the number of data points and m the number of vertices of the convex hull.

Value

A window (an object of class "owin").

Author(s)

Adrian Baddeley adrian@maths.uwa.edu.au http://www.maths.uwa.edu.au/~adrian/ and Rolf Turner rolf@math.unb.ca http://www.math.unb.ca/~rolf

References

Ripley, B.D. and Rasson, J.-P. (1977) Finding the edge of a Poisson forest. Journal of Applied Probability, 14, 483 – 491.

See Also

owin, as.owin

Examples


  ## Not run: 
  plot(owin())
  
## End(Not run)
  x <- runif(30)
  y <- runif(30)
  ## Not run: points(x,y)
  w <- ripras(x,y)
  ## Not run: 
  plot(w, box=FALSE)
  points(x,y)
  
## End(Not run)

[Package spatstat version 1.10-3 Index]