rlabel {spatstat}	R Documentation

Random Re-Labelling of Point Pattern

Description

Randomly allocates marks to a point pattern, or permutes the existing marks, or resamples from the existing marks.

Usage


   rlabel(X, labels=marks(X), permute=TRUE)

Arguments

X 	 Point pattern (object of class "ppp").

labels 	 Vector of values from which the new marks will be drawn at random. Defaults to the vector of existing marks.

permute 	 Logical value indicating whether to generate new marks by randomly permuting labels or by drawing a random sample with replacement.

Details

This very simple function allocates random marks to an existing point pattern X. It is useful for hypothesis testing purposes.

In the simplest case, the command rlabel(X) yields a point pattern obtained from X by randomly permuting the marks of the points.

If permute=TRUE, then labels should be a vector of length equal to the number of points in X. The result of rlabel will be a point pattern with locations given by X and marks given by a random permutation of labels (i.e. a random sample without replacement).

If permute=FALSE, then labels may be a vector of any length. The result of rlabel will be a point pattern with locations given by X and marks given by a random sample from labels (with replacement).

Value

A marked point pattern.

Author(s)

Adrian Baddeley adrian@maths.uwa.edu.au http://www.maths.uwa.edu.au/~adrian/ and Rolf Turner rolf@math.unb.ca http://www.math.unb.ca/~rolf

See Also

setmarks to assign arbitrary marks.

Examples


   data(amacrine)

   # Randomly permute the marks "on" and "off"
   # Result always has 142 "off" and 152 "on"
   Y <- rlabel(amacrine)

   # randomly allocate marks "on" and "off"
   # with probabilities p(off) = 0.48, p(on) = 0.52
   Y <- rlabel(amacrine, permute=FALSE)

   # randomly allocate marks "A" and "B" with equal probability
   data(cells)
   Y <- rlabel(cells, labels=factor(c("A", "B")), permute=FALSE)

[Package spatstat version 1.10-3 Index]