rotate.ppp {spatstat}	R Documentation

Rotate a Point Pattern

Description

Rotates a point pattern

Usage


 ## S3 method for class 'ppp':
 rotate(X, angle=pi/2, ...)

Arguments

X 	 A point pattern (object of class "ppp").

angle 	 Angle of rotation.

... 	 Ignored.

Details

The points of the pattern, and the window of observation, are rotated about the origin by the angle specified. Angles are measured in radians, anticlockwise. The default is to rotate the pattern 90 degrees anticlockwise. If the points carry marks, these are preserved.

Rotation of binary image masks is not yet implemented.

Value

Another object of class "ppp" representing the rotated point pattern.

Author(s)

Adrian Baddeley adrian@maths.uwa.edu.au http://www.maths.uwa.edu.au/~adrian/ and Rolf Turner rolf@math.unb.ca http://www.math.unb.ca/~rolf

See Also

ppp.object, rotate.owin

Examples


  data(cells)
  X <- rotate(cells, pi/3)
  ## Not run: 
  plot(X)
  
## End(Not run)

[Package spatstat version 1.10-3 Index]