rotate.psp {spatstat}	R Documentation

Rotate a Line Segment Pattern

Description

Rotates a line segment pattern

Usage


 ## S3 method for class 'psp':
 rotate(X, angle=pi/2, ...)

Arguments

X 	 A line segment pattern (object of class "psp").

angle 	 Angle of rotation.

... 	 Ignored.

Details

The line segments of the pattern, and the window of observation, are rotated about the origin by the angle specified. Angles are measured in radians, anticlockwise. The default is to rotate the pattern 90 degrees anticlockwise. If the line segments carry marks, these are preserved.

Rotation of binary image masks is not yet implemented.

Value

Another object of class "psp" representing the rotated line segment pattern.

Author(s)

Adrian Baddeley adrian@maths.uwa.edu.au http://www.maths.uwa.edu.au/~adrian/ and Rolf Turner rolf@math.unb.ca http://www.math.unb.ca/~rolf

See Also

psp.object, rotate.owin, rotate.ppp

Examples


  oldpar <- par(mfrow=c(2,1))
  X <- psp(runif(10), runif(10), runif(10), runif(10), window=owin())
  plot(X, main="original")
  Y <- rotate(X, pi/4)
  plot(Y, "rotated")
  par(oldpar)

[Package spatstat version 1.10-3 Index]