rpoisline {spatstat}	R Documentation

Generate Poisson Random Line Process

Description

Generate a random pattern of line segments obtained from the Poisson line process.

Usage


 rpoisline(lambda, win=owin())

Arguments

lambda 	 Intensity of the Poisson line process. A positive number.

win 	 Window in which to simulate the pattern. An object of class "owin" or something acceptable to as.owin. Currently, the window must be a rectangle.

Details

This algorithm generates a realisation of the uniform Poisson line process, and clips it to the window win.

Currently win must be a rectangle.

The argument lambda must be a positive number. It controls the intensity of the process. The expected number of lines intersecting a convex region of the plane is equal to lambda times the perimeter length of the region. The expected total length of the lines crossing a region of the plane is equal to lambda * pi times the area of the region.

Value

A line segment pattern (an object of class "psp").

Author(s)

Adrian Baddeley adrian@maths.uwa.edu.au http://www.maths.uwa.edu.au/~adrian/ and Rolf Turner rolf@math.unb.ca http://www.math.unb.ca/~rolf

See Also

psp

Examples


 # uniform Poisson line process with intensity 10,
 # clipped to the unit square
 rpoisline(10)

[Package spatstat version 1.10-3 Index]