rstrat {spatstat}	R Documentation

Simulate Stratified Random Point Pattern

Description

Generates a ``stratified random'' pattern of points in a window, by dividing the window into rectangular tiles and placing k random points independently in each tile.

Usage


 rstrat(win=square(1), nx, ny=nx, k = 1)

Arguments

win 	 A window. An object of class owin, or data in any format acceptable to as.owin().

nx 	 Number of tiles in each column.

ny 	 Number of tiles in each row.

k 	 Number of random points to generate in each tile.

Details

This function generates a random pattern of points in a ``stratified random'' sampling design. It can be useful for generating random spatial sampling points.

The bounding rectangle of win is divided into a regular nx * ny grid of rectangular tiles. In each tile, k random points are generated independently with a uniform distribution in that tile.

Some of these grid points may lie outside the window win: if they do, they are deleted.

The result is a point pattern inside the window win.

This function is useful in creating dummy points for quadrature schemes (see quadscheme) as well as in simulating random point patterns.

Value

A point pattern (object of class "ppp").

Author(s)

Adrian Baddeley adrian@maths.uwa.edu.au http://www.maths.uwa.edu.au/~adrian/ and Rolf Turner rolf@math.unb.ca http://www.math.unb.ca/~rolf

See Also

rsyst, runifpoint, quadscheme

Examples


  X <- rstrat(nx=10)
  plot(X)

  # polygonal boundary
  data(letterR)
  X <- rstrat(letterR, 5, 10, k=3)
  plot(X)

[Package spatstat version 1.10-3 Index]