rsyst {spatstat}	R Documentation

Simulate systematic random point pattern

Description

Generates a ``systematic random'' pattern of points in a window, consisting of a grid of equally-spaced points with a random common displacement.

Usage


 rsyst(win=square(1), nx, ny=nx, dx=NULL, dy=NULL)

Arguments

win 	 A window. An object of class owin, or data in any format acceptable to as.owin().

dx 	 Spacing of grid points in x direction. Incompatible with nx.

dy 	 Spacing of grid points in y direction. Incompatible with ny.

nx 	 Number of columns of grid points in the window. Incompatible with dx.

ny 	 Number of rows of grid points in the window. Incompatible with dy.

Details

This function generates a “systematic random” pattern of points in the window win. The pattern consists of a rectangular grid of points with a random common displacement.

The grid spacing is determined by the distances dx, dy if they are present. If they are absent, then the grid spacing is determined so that there will be nx columns and ny rows of grid points in the bounding rectangle of win.

The grid is then given a random displacement (the common displacement of the grid points is a uniformly distributed random vector in the tile of dimensions dx, dy).

Some of the resulting grid points may lie outside the window win: if they do, they are deleted. The result is a point pattern inside the window win.

This function is useful in creating dummy points for quadrature schemes (see quadscheme) as well as in simulating random point patterns.

Value

A point pattern (object of class "ppp").

Author(s)

Adrian Baddeley adrian@maths.uwa.edu.au http://www.maths.uwa.edu.au/~adrian/ and Rolf Turner rolf@math.unb.ca http://www.math.unb.ca/~rolf

See Also

rstrat, runifpoint, quadscheme

Examples


  X <- rsyst(nx=10)
  plot(X)

  # polygonal boundary
  data(letterR)
  X <- rsyst(letterR, 5, 10)
  plot(X)

[Package spatstat version 1.10-3 Index]