scanpp {spatstat}	R Documentation

Read Point Pattern From Data File

Description

Reads a point pattern dataset from a text file.

Usage


   scanpp(filename, window, header=TRUE, dir="", multitype=FALSE) 

Arguments

filename 	 String name of the file containing the coordinates of the points in the point pattern, and their marks if any.

window 	 Window for the point pattern. An object of class "owin".

header 	 Logical flag indicating whether the first line of the file contains headings for the columns. Passed to read.table.

dir 	 String containing the path name of the directory in which filename is to be found. Default is the current directory.

multitype 	 Logical flag indicating whether marks are to be interpreted as a factor (multitype = TRUE) or as numerical values (multitype = FALSE).

Details

This simple function reads a point pattern dataset from a file containing the cartesian coordinates of its points, and optionally the mark values for these points.

The file identified by filename in directory dir should be a text file that can be read using read.table. Thus, each line of the file (except possibly the first line) contains data for one point in the point pattern. Data are arranged in columns. There should be either two columns (for an unmarked point pattern) or three columns (for a marked point pattern).

If header=FALSE then the first two columns of data will be interpreted as the x and y coordinates of points. A third column, if present, will be interpreted as containing the marks for these points. The marks will be converted to a factor if multitype = TRUE.

If header=TRUE then the first line of the file should contain string names for each of the columns of data. If there are columns named x and y then these will be taken as the cartesian coordinates, and any remaining column will be taken as the marks. If there are no columns named x and y then the first and second columns will be taken as the cartesian coordinates.

Note that there is intentionally no default for window. The window of observation should be specified. If you really need to estimate the window, use the Ripley-Rasson estimator ripras.

Value

A point pattern (an object of class "ppp", see ppp.object).

Author(s)

Adrian Baddeley adrian@maths.uwa.edu.au http://www.maths.uwa.edu.au/~adrian/ and Rolf Turner rolf@math.unb.ca http://www.math.unb.ca/~rolf

See Also

ppp.object, ppp, as.ppp, ripras

[Package spatstat version 1.10-3 Index]