setcov {spatstat}	R Documentation

Set Covariance of a Window

Description

Computes the set covariance function of a window.

Usage


 setcov(W, ...)

Arguments

W 	 A window (object of class "owin".

... 	 Optional arguments passed to as.mask to control the pixel resolution.

Details

The set covariance function of a region W in the plane is the function C(v) defined for each vector v as the area of the intersection between W and W+v, where W+v is the set obtained by shifting (translating) W by v.

We may interpret C(v) as the area of the set of all points x in W such that x+v also lies in W.

This command computes a discretised approximation to the set covariance function of any plane region W represented as a window object (of class "owin", see owin.object). The return value is a pixel image (object of class "im") whose greyscale values are values of the set covariance function.

The set covariance is computed using the Fast Fourier Transform, unless W is a rectangle, when an exact formula is used.

Value

A pixel image (an object of class "im") representing the set covariance function of W.

Author(s)

Adrian Baddeley adrian@maths.uwa.edu.au http://www.maths.uwa.edu.au/~adrian/ and Rolf Turner rolf@math.unb.ca http://www.math.unb.ca/~rolf

See Also

owin, as.owin, erode.owin

Examples


  w <- owin(c(0,1),c(0,1))
  v <- setcov(w)
  plot(v)

[Package spatstat version 1.10-3 Index]