shift.ppp {spatstat}	R Documentation

Apply Vector Translation To Point Pattern

Description

Applies a vector shift to a point pattern.

Usage


 ## S3 method for class 'ppp':
 shift(X, vec=c(0,0), ...)

Arguments

X 	 Point pattern (object of class "ppp").

vec 	 Vector of length 2 representing a translation.

... 	 Ignored

Details

The point pattern, and its window, are translated by the vector vec.

This is a method for the generic function shift.

Value

Another point pattern (of class "ppp") representing the result of applying the vector shift.

Author(s)

Adrian Baddeley adrian@maths.uwa.edu.au http://www.maths.uwa.edu.au/~adrian/ and Rolf Turner rolf@math.unb.ca http://www.math.unb.ca/~rolf

See Also

shift, shift.owin, rotate, affine

Examples


  data(cells)
  X <- shift(cells, c(2,3))
  ## Not run: 
  plot(X)
  # no discernible difference except coordinates are different
  
## End(Not run)
  Y <- superimpose(cells, shift(cells, c(0.1,0.1)))
  ## Not run: plot(Y)

[Package spatstat version 1.10-3 Index]