simdat {spatstat}	R Documentation

Simulated Point Pattern

Description

This point pattern data set was simulated (using the Metropolis-Hastings algorithm) from a model fitted to the Numata Japanese black pine data set referred to in Baddeley and Turner (2000).

Usage

data(simdat)

Format

An object of class "ppp" in a square window of size 10 by 10 units.

See ppp.object for details of the format of a point pattern object.

Source

Rolf Turner rolf@math.unb.ca

References

Baddeley, A. and Turner, R. (2000) Practical maximum pseudolikelihood for spatial point patterns. Australian and New Zealand Journal of Statistics 42, 283–322.

[Package spatstat version 1.10-3 Index]