spatstat.family {spatstat}	R Documentation

Names of All Packages in the Spatstat Family

Description

Provides the names of all the packages belonging to the spatstat family of packages.

Usage

  spatstat.family(subpackages=TRUE, extensions=FALSE)

Arguments

subpackages
Logical value specifying whether to include sub-packages.

extensions
Logical value specifying whether to include extension packages.

Details

This function returns a character vector containing the names of the packages that belong to the spatstat family.

By default, only the sub-packages are listed, and not the extension packages.

A “sub-package” is a package which is implicitly loaded or imported when the command library(spatstat) is issued. Currently the sub-packages are:

spatstat.utils

spatstat.data

spatstat.univar

spatstat.sparse

spatstat.geom

spatstat.random

spatstat.explore

spatstat.model

spatstat.linnet

spatstat

An “extension package” is a package which must be loaded explicitly. The extension packages are:

spatstat.gui

spatstat.local

spatstat.Knet

Value

Character vector of package names.

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au, Rolf Turner rolfturner@posteo.net and Ege Rubak rubak@math.aau.dk.

See Also

latest.news

[Package spatstat version 3.3-1 Index]