spatstat.options {spatstat}	R Documentation

Internal Options in Spatstat Package

Description

Allows the user to examine and reset the values of global parameters which control actions in the spatstat package.

Usage


  spatstat.options(...)

Arguments

... 	 Either empty, or a succession of parameter names in quotes, or a succession of name=value pairs. See below for the parameter names.

Details

This function allows the user to examine and reset the values of global parameters which control actions in the spatstat package. It is analogous to the system function options.

The global parameters are:

npixel

Controls the number of pixels in the discrete approximation of an irregular window by a binary pixel image as created by as.mask. Either an integer or a pair of integers giving the number of pixels in the x and y directions.

maxedgewt

Edge correction weights will be trimmed so as not to exceed this value. This applies to the weights computed by edge.Trans or edge.Ripley and used in Kest and its relatives.

par.binary

List of arguments to be passed to the function image when displaying a binary image mask (in plot.owin or plot.ppp). Typically used to reset the colours of foreground and background.

par.persp

List of arguments to be passed to the function persp when displaying a real-valued image, such as the fitted surfaces in plot.ppm.

par.points

List of arguments controlling the plotting of point patterns by plot.ppp.

par.contour

List of arguments controlling contour plots of pixel images by contour.im.

If no arguments are given, the current values of all parameters are returned, in a list.

If one parameter name is given, the current value of this parameter is returned (not in a list, just the value).

If several parameter names are given, the current values of these parameters are returned, in a list.

If name=value pairs are given, the named parameters are reset to the given values, and the previous values of these parameters are returned, in a list.

Value

Either a list of parameters and their values, or a single value. See Details.

Author(s)

Adrian Baddeley adrian@maths.uwa.edu.au http://www.maths.uwa.edu.au/~adrian/ and Rolf Turner rolf@math.unb.ca http://www.math.unb.ca/~rolf

See Also

options

Examples


  spatstat.options()

  spatstat.options("npixel")
  spatstat.options(npixel=150)
  spatstat.options(npixel=c(100,200))

  spatstat.options(par.binary=list(col=grey(c(0.5,1))))

  spatstat.options(par.persp=list(theta=-30,phi=40,d=4))
  # see help(persp.default) for other options


[Package spatstat version 1.10-3 Index]