split.ppp {spatstat}	R Documentation

Divide Point Pattern into Sub-patterns

Description

Divides a point pattern into several sub-patterns, according to their marks, or according to any user-specified grouping.

Usage


  ## S3 method for class 'ppp':
  split(x, f = marks(x), drop=FALSE, un=NULL, ...)
  ## S3 method for class 'ppp':
  split(x, f = marks(x), drop=FALSE, un=missing(f), ...) <- value

Arguments

x 	 A two-dimensional point pattern. An object of class "ppp".

f 	 Factor determining the grouping.

drop 	 Logical. Determines whether empty groups will be deleted.

un 	 Logical. Determines whether subpatterns will be unmarked (i.e. whether marks will be removed from the points in each subpattern).

... 	 Other arguments are ignored.

value 	 List of point patterns.

Details

The function split.ppp divides up the points of the point pattern x into several sub-patterns according to the levels of the factor f. The result is a list of point patterns, one for each level of f.

If f is present, it must be a factor, and its length must equal the number of points in x. The levels of f determine the destination of each point in x. The ith point of x will be placed in the sub-pattern split.ppp(x)$l where l = f[i].

If f is missing, then x must be a multitype point pattern (a marked point pattern whose marks vector is a factor). Then the effect is that the points of each type are separated into different point patterns.

Some of the sub-patterns created by the split may be empty. If drop=TRUE, then empty sub-patterns will be deleted from the list. If drop=FALSE then they are retained.

The argument un determines how to handle marks in the case where x is a marked point pattern. If un=TRUE then the marks of the points will be discarded when they are split into groups, while if un=FALSE then the marks will be retained.

The result of split.ppp has class "splitppp" and can be plotted using plot.splitppp.

The assignment function split<-.ppp updates the point pattern x so that it satisfies split(x, f, drop, un) = value. The argument value is expected to be a list of point patterns, one for each level of f. These point patterns are expected to be compatible in the sense that they all have the same window, and either they are all unmarked or they all have marks of the same kind.

Splitting can also be undone by the function superimpose.

Value

The value of split.ppp is a list of point patterns. The components of the list are named by the levels of f.
The assignment form split<-.ppp returns the updated point pattern x.

Author(s)

Adrian Baddeley adrian@maths.uwa.edu.au http://www.maths.uwa.edu.au/~adrian/ and Rolf Turner rolf@math.unb.ca http://www.math.unb.ca/~rolf

See Also

cut.ppp, plot.splitppp, superimpose, ppp.object

Examples



# Multitype point pattern: separate into types
 data(amacrine)
 u <- split(amacrine)

# plot them
 plot(split(amacrine))

# the following are equivalent:
 amon <- split(amacrine)$on
 amon <- unmark(amacrine[amacrine$marks == "on"])
   
# the following are equivalent:
 amon <- split(amacrine, un=FALSE)$on
 amon <- amacrine[amacrine$marks == "on"]
   
# Scramble the locations of the 'on' cells
 u <- split(amacrine)
 u$on <- runifpoint(amon$n, amon$window)
 split(amacrine) <- u

# Point pattern with continuous marks
 data(longleaf)
 
 # cut the range of tree diameters into three intervals
 long3 <- cut.ppp(longleaf, 3)
 # now split them
 long3split <- split(long3)

# Unmarked point pattern
  data(swedishpines)
# cut & split according to nearest neighbour distance
  f <- cut(nndist(swedishpines), 3)
  u <- split(swedishpines, f)

[Package spatstat version 1.10-3 Index]