spokes {spatstat}	R Documentation

Spokes pattern of dummy points

Description

Generates a pattern of dummy points in a window, given a data point pattern. The dummy points lie on the radii of circles emanating from each data point.

Usage


 spokes(x, y, nrad = 3, nper = 3, fctr = 1.5, Mdefault = 1)

Arguments

x 	 Vector of x coordinates of data points, or a list with components x and y, or a point pattern (an object of class ppp).

y 	 Vector of y coordinates of data points. Ignored unless x is a vector.

nrad 	 Number of radii emanating from each data point.

nper 	 Number of dummy points per radius.

fctr 	 Scale factor. Length of largest spoke radius is fctr * M where M is the mean nearest neighbour distance for the data points.

Mdefault 	 Value of M to be used if x has length 1.

Details

This function is useful in creating dummy points for quadrature schemes (see quadscheme).

Given the data points, the function creates a collection of nrad * nper * length(x) dummy points.

Around each data point (x[i],y[i]) there are nrad * nper dummy points, lying on nrad radii emanating from (x[i],y[i]), with nper dummy points equally spaced along each radius.

The (equal) spacing of dummy points along each radius is controlled by the factor fctr. The distance from a data point to the furthest of its associated dummy points is fctr * M where M is the mean nearest neighbour distance for the data points.

If there is only one data point the nearest neighbour distance is infinite, so the value Mdefault will be used in place of M.

If x is a point pattern, then the value returned is also a point pattern, which is clipped to the window of x. Hence there may be fewer than nrad * nper * length(x) dummy points in the pattern returned.

Value

If argument x is a point pattern, a point pattern with window equal to that of x. Otherwise a list with two components x and y. In either case the components x and y of the value are numeric vectors giving the coordinates of the dummy points.

Author(s)

Adrian Baddeley adrian@maths.uwa.edu.au http://www.maths.uwa.edu.au/~adrian/ and Rolf Turner rolf@math.unb.ca http://www.math.unb.ca/~rolf

See Also

quad.object, quadscheme, inside.owin, gridcentres, stratrand

Examples


  dat <- runifrect(10)
  ## Not run: 
  plot(dat)
  
## End(Not run)
  dum <- spokes(dat$x, dat$y)
  ## Not run: 
  points(dum$x, dum$y, pch=".")
  
## End(Not run)
  Q <- quadscheme(dat, dum)

[Package spatstat version 1.10-3 Index]