square {spatstat}	R Documentation

Square Window

Description

Creates a square window

Usage


 square(r=1)
 unit.square()

Arguments

r 	 The side length of the square.

Details

square is a shortcut for creating a window object representing the square [0,r] * [0,r]. It is equivalent to the command owin(c(0,r),c(0,r)).

unit.square creates the unit square [0,1] * [0,1]. It is equivalent to square(1) or square() or owin(c(0,1),c(0,1)).

These commands are included mainly to improve the readability of some code.

Value

An object of class "owin" (see owin.object) specifying a window.

Author(s)

Adrian Baddeley adrian@maths.uwa.edu.au http://www.maths.uwa.edu.au/~adrian/ and Rolf Turner rolf@math.unb.ca http://www.math.unb.ca/~rolf

See Also

owin.object, owin

Examples


 W <- square(10)

[Package spatstat version 1.10-3 Index]