subset.fasp {spatstat}	R Documentation

Extract Subset of Function Array

Description

Extract a subset of a function array (an object of class "fasp").

Usage


  ## S3 method for class 'fasp':
  x[I, J, drop, ...]

Arguments

x 	 A function array. An object of class "fasp".

I 	 any valid expression for a subset of the row indices of the array.

J 	 any valid expression for a subset of the column indices of the array.

drop, ... 	 Ignored.

Details

A function array can be regarded as a matrix whose entries are functions. See fasp.object for an explanation of function arrays.

This routine extracts a sub-array according to the usual conventions for matrix indexing.

Value

A function array (of class "fasp").

Author(s)

Adrian Baddeley adrian@maths.uwa.edu.au http://www.maths.uwa.edu.au/~adrian/ and Rolf Turner rolf@math.unb.ca http://www.math.unb.ca/~rolf

See Also

fasp.object

Examples


 # Lansing woods data - multitype points with 6 types
 data(lansing)
 
 # compute 6 x 6 array of all cross-type K functions
 a <- alltypes(lansing, "K")

 # extract first three marks only
 b <- a[1:3,1:3]
 ## Not run: plot(b)
 # subset pertaining to hickories
 h <- a[levels(lansing$marks) == "hickory", ]
 ## Not run: plot(h)

[Package spatstat version 1.10-3 Index]