subset.fv {spatstat}	R Documentation

Extract Subset of Function Values

Description

Extract a subset of an object of class "fv".

Usage


  ## S3 method for class 'fv':
  x[i, j, ..., drop=FALSE]

Arguments

x 	 a function value object, of class "fv" (see fv.object). Essentially a data frame.

i 	 any appropriate subset index. Selects a subset of the rows of the data frame, i.e. a subset of the domain of the function(s) represented by x.

j 	 any appropriate subset index for the columns of the data frame. Selects some of the functions present in x.

drop, ... 	 Ignored.

Details

This is a method for "[" for the class "fv". It is very similar to [.data.frame except for a few extra checks on the sanity of the result.

Value

A function value object (of class "fv").

Author(s)

Adrian Baddeley adrian@maths.uwa.edu.au http://www.maths.uwa.edu.au/~adrian/ and Rolf Turner rolf@math.unb.ca http://www.math.unb.ca/~rolf

See Also

fv.object

Examples


 data(cells)
 K <- Kest(cells)

 # discard the estimates of K(r) for r  > 0.1
 Ksub <- K[K$r <= 0.1, ]

 # discard the border method estimator
 Ksub <- K[ , names(K) != "border"]
 

[Package spatstat version 1.10-3 Index]