summary.im {spatstat}	R Documentation

Summarizing a Pixel Image

Description

summary method for class "im".

Usage


  ## S3 method for class 'im':
  summary(object, ...)
  ## S3 method for class 'im':
  print.summary(x, ...)

Arguments

object 	 A pixel image.

... 	 Ignored.

x 	 Object of class "summary.im" as returned by summary.im.

Details

This is a method for the generic summary for the class "im". An object of class "im" describes a pixel image. See im.object) for details of this class.

summary.im extracts information about the pixel image, and print.summary.im prints this information in a comprehensible format.

In normal usage, print.summary.im is invoked implicitly when the user calls summary.im without assigning its value to anything. See the examples.

The information extracted by summary.im includes

range

The range of the image values.

mean

The mean of the image values.

integral

The ``integral'' of the image values, calculated as the sum of the image values multiplied by the area of one pixel.

dim

The dimensions of the pixel array: dim[1] is the number of rows in the array, corresponding to the y coordinate.

Value

summary.im returns an object of class "summary.im", while print.summary.im returns NULL.

Author(s)

Adrian Baddeley adrian@maths.uwa.edu.au http://www.maths.uwa.edu.au/~adrian/ and Rolf Turner rolf@math.unb.ca http://www.math.unb.ca/~rolf

Examples


  # make an image
  X <- as.im(function(x,y) {x^2}, unit.square())
  # summarize it
  summary(X)
  # save the summary
  s <- summary(X)
  # print it
  print(X)
  s
  # extract stuff
  X$dim
  X$range
  X$integral

[Package spatstat version 1.10-3 Index]