summary.owin {spatstat}	R Documentation

Summary of a Spatial Window

Description

Prints a useful description of a window object.

Usage


  ## S3 method for class 'owin':
  summary(object, ...)

Arguments

object 	 Window (object of class "owin").

... 	 Ignored.

Details

A useful description of the window object is printed.

This is a method for the generic function summary.

Author(s)

Adrian Baddeley adrian@maths.uwa.edu.au http://www.maths.uwa.edu.au/~adrian/ and Rolf Turner rolf@math.unb.ca http://www.math.unb.ca/~rolf

See Also

summary, summary.ppp, print.owin

Examples


  summary(owin())  # the unit square

  data(demopat)
  W <- demopat$window  # weird polygonal window
  summary(W)           # describes it

  summary(as.mask(W))  # demonstrates current pixel resolution

[Package spatstat version 1.10-3 Index]