summary.ppm {spatstat}	R Documentation

Summarizing a Fitted Point Process Model

Description

summary method for class "ppm".

Usage


  ## S3 method for class 'ppm':
  summary(object, ..., quick=FALSE)
  ## S3 method for class 'ppm':
  print.summary(x, ...)

Arguments

object 	 A fitted point process model.

... 	 Ignored.

quick 	 Logical flag controlling the scope of the summary.

x 	 Object of class "summary.ppm" as returned by summary.ppm.

Details

This is a method for the generic summary for the class "ppm". An object of class "ppm" describes a fitted point process model. See ppm.object) for details of this class.

summary.ppm extracts information about the type of model that has been fitted, the data to which the model was fitted, and the values of the fitted coefficients. (If quick=TRUE then only the information about the type of model is extracted.)

print.summary.ppm prints this information in a comprehensible format.

In normal usage, print.summary.ppm is invoked implicitly when the user calls summary.ppm without assigning its value to anything. See the examples.

Value

summary.ppm returns an object of class "summary.ppm", while print.summary.ppm returns NULL.

Author(s)

Adrian Baddeley adrian@maths.uwa.edu.au http://www.maths.uwa.edu.au/~adrian/ and Rolf Turner rolf@math.unb.ca http://www.math.unb.ca/~rolf

Examples


  # invent some data
  X <- rpoispp(42)
  # fit a model to it
  fit <- ppm(X, ~x, Strauss(r=0.1))
  # summarize the fitted model
  summary(fit)
  # `quick' option
  summary(fit, quick=TRUE)

  # save the full summary
  s <- summary(fit)
  # print it
  print(s)
  s
  # extract stuff
  names(s)
  s$args$correction
  s$name
  s$trend$value

  ## Not run: 
  # multitype pattern
  data(demopat)
  fit <- ppm(demopat, ~marks, Poisson())
  summary(fit)
  
## End(Not run)

[Package spatstat version 1.10-3 Index]