summary.ppp {spatstat}	R Documentation

Summary of a Point Pattern Dataset

Description

Prints a useful summary of a point pattern dataset.

Usage


 ## S3 method for class 'ppp':
 summary(object, ..., checkdup=TRUE)

Arguments

object 	 Point pattern (object of class "ppp").

... 	 Ignored.

checkdup 	 Logical value indicating whether to check for the presence of duplicate points.

Details

A useful summary of the point pattern object is printed.

This is a method for the generic function summary.

If checkdup=TRUE, the pattern will be checked for the presence of dublicate points, using duplicated.ppp. This can be time-consuming if the pattern contains many points, so the checking can be disabled by setting checkdup=FALSE.

Author(s)

Adrian Baddeley adrian@maths.uwa.edu.au http://www.maths.uwa.edu.au/~adrian/ and Rolf Turner rolf@math.unb.ca http://www.math.unb.ca/~rolf

See Also

summary, summary.owin, print.ppp

Examples


  data(cells)      # plain vanilla point pattern
  summary(cells)  

  data(lansing)    # multitype point pattern
  
  summary(lansing) # tabulates frequencies of each mark
  
  data(longleaf)    # numeric marks
  
  summary(longleaf) # prints summary.default(x$marks)

  data(demopat)     # weird polygonal window
  summary(demopat)  # describes it

[Package spatstat version 1.10-3 Index]