summary.quad {spatstat}	R Documentation

Summarizing a Quadrature Scheme

Description

summary method for class "quad".

Usage


  ## S3 method for class 'quad':
  summary(object, ..., checkdup=FALSE)
  ## S3 method for class 'quad':
  print.summary(x, ..., dp=3)

Arguments

object 	 A quadrature scheme.

... 	 Ignored.

checkdup 	 Logical value indicating whether to test for duplicated points.

dp 	 Number of significant digits to print.

x 	 Object of class "summary.quad" returned by summary.quad.

Details

This is a method for the generic summary for the class "quad". An object of class "quad" describes a quadrature scheme, used to fit a point process model. See quad.object) for details of this class.

summary.quad extracts information about the quadrature scheme, and print.summary.quad prints this information in a comprehensible format.

In normal usage, print.summary.quad is invoked implicitly when the user calls summary.quad without assigning its value to anything. See the examples.

Value

summary.quad returns an object of class "summary.quad", while print.summary.quad returns NULL.

Author(s)

Adrian Baddeley adrian@maths.uwa.edu.au http://www.maths.uwa.edu.au/~adrian/ and Rolf Turner rolf@math.unb.ca http://www.math.unb.ca/~rolf

Examples


  # make a quadrature scheme
  Q <- quadscheme(rpoispp(42))
  # summarize it
  summary(Q)
  # save the summary
  s <- summary(Q)
  # print it
  print(s)
  s
  # extract total quadrature weight
  s$w$all$sum

[Package spatstat version 1.10-3 Index]