union.quad {spatstat}	R Documentation

Union of Data and Dummy Points

Description

Combines the data and dummy points of a quadrature scheme into a single point pattern.

Usage


 union.quad(Q)

Arguments

Q 	 A quadrature scheme (an object of class "quad").

Details

The argument Q should be a quadrature scheme (an object of class "quad", see quad.object for details).

This function combines the data and dummy points of Q into a single point pattern. If either the data or the dummy points are marked, the result is a marked point pattern.

The function as.ppp will perform the same task.

Value

A point pattern (of class "ppp").

Author(s)

Adrian Baddeley adrian@maths.uwa.edu.au http://www.maths.uwa.edu.au/~adrian/ and Rolf Turner rolf@math.unb.ca http://www.math.unb.ca/~rolf

See Also

quad.object, as.ppp

Examples


  data(simdat)
  Q <- quadscheme(simdat, default.dummy(simdat))
  U <- union.quad(Q)
  ## Not run: plot(U)
  # equivalent:
  U <- as.ppp(Q)

[Package spatstat version 1.10-3 Index]