unique.ppp {spatstat}	R Documentation

Extract Unique Points from a Spatial Point Pattern

Description

Removes any points that are identical to other points in a spatial point pattern.

Usage


 ## S3 method for class 'ppp':
 unique(x, ...)

Arguments

x 	 A spatial point pattern (object of class "ppp").

... 	 Ignored.

Details

This is a method for the generic function unique for point pattern datasets (of class "ppp", see ppp.object).

Two points in a point pattern are deemed to be identical if their x,y coordinates are the same, and their marks are also the same (if they carry marks). The Examples section illustrates how it is possible for a point pattern to contain a pair of identical points.

This function removes duplicate points in x, and returns a point pattern.

Value

Another point pattern object.

Author(s)

Adrian Baddeley adrian@maths.uwa.edu.au http://www.maths.uwa.edu.au/~adrian/ and Rolf Turner rolf@math.unb.ca http://www.math.unb.ca/~rolf

See Also

ppp.object, duplicated.ppp, multiplicity.ppp

Examples


   X <- ppp(c(1,1,0.5), c(2,2,1), window=square(3))
   unique(X)

[Package spatstat version 1.10-3 Index]