units {spatstat}	R Documentation

Name for Unit of Length

Description

Inspect or change the name of the unit of length in a spatial dataset.

Usage


units(x)
## S3 method for class 'ppp':
units(x)
## S3 method for class 'psp':
units(x)
## S3 method for class 'owin':
units(x)
## S3 method for class 'im':
units(x)
units(x) <- value
## S3 method for class 'ppp':
units(x) <- value
## S3 method for class 'psp':
units(x) <- value
## S3 method for class 'owin':
units(x) <- value
## S3 method for class 'im':
units(x) <- value

Arguments

x 	 A spatial dataset. Either a point pattern (object of class "ppp"), a line segment pattern (object of class "psp"), a window (object of class "owin") or a pixel image (object of class "im").

value 	 Name of the unit of length. See Details.

Details

Spatial datasets in the spatstat package may include the name of the unit of length. This name is used when printing or plotting the dataset, and in some other applications.

units(x) extracts this name, and units(x) <- value sets the name to value.

A valid name is either

a single character string
a vector of two character strings giving the singular and plural forms of the unit name
a list of length 3, containing two character strings giving the singular and plural forms of the basic unit, and a number specifying the multiple of this unit.

Value

The return value of units is an object of class "units" containing the name of the unit of length in x. There are methods for print and summary.

Author(s)

Adrian Baddeley adrian@maths.uwa.edu.au http://www.maths.uwa.edu.au/~adrian/ and Rolf Turner rolf@math.unb.ca http://www.math.unb.ca/~rolf

See Also

owin, ppp

Examples


  X <- runifpoint(20)

  # if the unit of length is 1 metre:
  units(X) <- c("metre", "metres")

  # if the unit of length is 6 inches:
  units(X) <- list("inch", "inches", 6)

[Package spatstat version 1.10-3 Index]