unmark {spatstat}	R Documentation

Remove Marks from a Marked Point Pattern

Description

Removes the mark information from a marked point pattern

Usage


 unmark(X)

Arguments

X 	 A point pattern. Either an object of class "ppp" or anything acceptable to as.ppp.

Details

This is simply a convenience for those occasions in which we want to remove the marks from a marked point pattern.

Value

A point pattern (of class "ppp") with any mark information deleted.

Author(s)

Adrian Baddeley adrian@maths.uwa.edu.au http://www.maths.uwa.edu.au/~adrian/ and Rolf Turner rolf@math.unb.ca http://www.math.unb.ca/~rolf

See Also

ppp.object, as.ppp

Examples


  data(lansing)
  hicks <- lansing[lansing$marks == "hickory", ]
  ## Not run: 
  plot(hicks)  # still a marked point pattern, but only 1 value of marks
  plot(unmark(hicks)) # unmarked
  
## End(Not run)

[Package spatstat version 1.10-3 Index]